Crate objc2_core_foundation

Bindings to the CoreFoundation framework

See Apple’s docs and the general docs on framework crates for more information.

Structs

CFAllocator

Apple’s documentation

CFAllocatorContext

Apple’s documentation

CFArray

This is the type of a reference to immutable CFArrays.

CFArrayCallBacks

Apple’s documentation

CFArrayIntoIter

A retained iterator over the items of an array.

CFArrayIter

An iterator over retained objects of an array.

CFArrayIterUnchecked

An iterator over raw items of an array.

CFBoolean

Apple’s documentation

CFComparisonResult

Apple’s documentation

CFData

Apple’s documentation

CFDataSearchFlags

Apple’s documentation

CFDictionary

This is the type of a reference to immutable CFDictionarys.

CFDictionaryKeyCallBacks

Apple’s documentation

CFDictionaryValueCallBacks

Structure containing the callbacks for values of a CFDictionary. Field: version The version number of the structure type being passed in as a parameter to the CFDictionary creation functions. This structure is version 0. Field: retain The callback used to add a retain for the dictionary on values as they are put into the dictionary. This callback returns the value to use as the value in the dictionary, which is usually the value parameter passed to this callback, but may be a different value if a different value should be added to the dictionary. The dictionary’s allocator is passed as the first argument. Field: release The callback used to remove a retain previously added for the dictionary from values as they are removed from the dictionary. The dictionary’s allocator is passed as the first argument. Field: copyDescription The callback used to create a descriptive string representation of each value in the dictionary. This is used by the CFCopyDescription() function. Field: equal The callback used to compare values in the dictionary for equality in some operations.

CFError

This is the type of a reference to CFErrors. CFErrorRef is toll-free bridged with NSError.

CFMutableArray

This is the type of a reference to mutable CFArrays.

CFMutableData

Apple’s documentation

CFMutableDictionary

This is the type of a reference to mutable CFDictionarys.

CFMutableString

Apple’s documentation

CFNull

Apple’s documentation

CFNumber

Apple’s documentation

CFNumberType

Apple’s documentation

CFRange

Apple’s documentation

CFRetained

A reference counted pointer type for CoreFoundation types.

CFRunLoop

Apple’s documentation

CFRunLoopActivity

Apple’s documentation

CFRunLoopObserver

Apple’s documentation

CFRunLoopObserverContext

Apple’s documentation

CFRunLoopRunResult

Apple’s documentation

CFRunLoopSource

Apple’s documentation

CFRunLoopSourceContext

Apple’s documentation

CFRunLoopTimer

Apple’s documentation

CFRunLoopTimerContext

Apple’s documentation

CFString

Apple’s documentation

CFStringBuiltInEncodings

Apple’s documentation

CFStringCompareFlags

Apple’s documentation

CFStringInlineBuffer

Apple’s documentation

CFStringNormalizationForm

This is the type of Unicode normalization forms as described in Unicode Technical Report #15. To normalize for use with file system calls, use CFStringGetFileSystemRepresentation().

CFType

An instance of a Core Foundation type.

CFURL

Apple’s documentation

CFURLBookmarkCreationOptions

Apple’s documentation

CFURLBookmarkResolutionOptions

Apple’s documentation

CFURLComponentType

Apple’s documentation

CFURLPathStyle

Apple’s documentation

CFUUID

Apple’s documentation

CFUUIDBytes

Apple’s documentation

IUnknownVTbl

Apple’s documentation

Constants

kCFCoreFoundationVersionNumber10_0

Apple’s documentation

kCFCoreFoundationVersionNumber10_0_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_5

Apple’s documentation

kCFCoreFoundationVersionNumber10_6

Apple’s documentation

kCFCoreFoundationVersionNumber10_7

Apple’s documentation

kCFCoreFoundationVersionNumber10_8

Apple’s documentation

kCFCoreFoundationVersionNumber10_9

Apple’s documentation

kCFCoreFoundationVersionNumber10_1_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_1_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_1_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_1_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_2_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_2_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_2_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_2_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_2_5

Apple’s documentation

kCFCoreFoundationVersionNumber10_2_6

Apple’s documentation

kCFCoreFoundationVersionNumber10_2_7

Apple’s documentation

kCFCoreFoundationVersionNumber10_2_8

Apple’s documentation

kCFCoreFoundationVersionNumber10_3_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_3_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_3_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_3_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_3_5

Apple’s documentation

kCFCoreFoundationVersionNumber10_3_6

Apple’s documentation

kCFCoreFoundationVersionNumber10_3_7

Apple’s documentation

kCFCoreFoundationVersionNumber10_3_8

Apple’s documentation

kCFCoreFoundationVersionNumber10_3_9

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_7

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_8

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_9

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_4_Intel

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_4_PowerPC

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_5_Intel

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_5_PowerPC

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_6_Intel

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_6_PowerPC

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_10

Apple’s documentation

kCFCoreFoundationVersionNumber10_4_11

Apple’s documentation

kCFCoreFoundationVersionNumber10_5_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_5_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_5_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_5_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_5_5

Apple’s documentation

kCFCoreFoundationVersionNumber10_5_6

Apple’s documentation

kCFCoreFoundationVersionNumber10_5_7

Apple’s documentation

kCFCoreFoundationVersionNumber10_5_8

Apple’s documentation

kCFCoreFoundationVersionNumber10_6_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_6_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_6_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_6_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_6_5

Apple’s documentation

kCFCoreFoundationVersionNumber10_6_6

Apple’s documentation

kCFCoreFoundationVersionNumber10_6_7

Apple’s documentation

kCFCoreFoundationVersionNumber10_6_8

Apple’s documentation

kCFCoreFoundationVersionNumber10_7_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_7_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_7_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_7_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_7_5

Apple’s documentation

kCFCoreFoundationVersionNumber10_8_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_8_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_8_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_8_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_9_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_9_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_10

Apple’s documentation

kCFCoreFoundationVersionNumber10_11

Apple’s documentation

kCFCoreFoundationVersionNumber10_10_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_10_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_10_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_10_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_10_5

Apple’s documentation

kCFCoreFoundationVersionNumber10_10_Max

Apple’s documentation

kCFCoreFoundationVersionNumber10_11_1

Apple’s documentation

kCFCoreFoundationVersionNumber10_11_2

Apple’s documentation

kCFCoreFoundationVersionNumber10_11_3

Apple’s documentation

kCFCoreFoundationVersionNumber10_11_4

Apple’s documentation

kCFCoreFoundationVersionNumber10_11_Max

Apple’s documentation

kCFStringEncodingInvalidId

Apple’s documentation

Statics

kCFAllocatorDefault⚠

Apple’s documentation

kCFAllocatorMalloc⚠

Apple’s documentation

kCFAllocatorMallocZone⚠

Apple’s documentation

kCFAllocatorNull⚠

Apple’s documentation

kCFAllocatorSystemDefault⚠

Apple’s documentation

kCFAllocatorUseContext⚠

Apple’s documentation

kCFBooleanFalse⚠

Apple’s documentation

kCFBooleanTrue⚠

Apple’s documentation

kCFCopyStringDictionaryKeyCallBacks⚠

Predefined CFDictionaryKeyCallBacks structure containing a set of callbacks appropriate for use when the keys of a CFDictionary are all CFStrings, which may be mutable and need to be copied in order to serve as constant keys for the values in the dictionary.

kCFCoreFoundationVersionNumber⚠

Apple’s documentation

kCFErrorDescriptionKey⚠

Apple’s documentation

kCFErrorDomainCocoa⚠

Apple’s documentation

kCFErrorDomainMach⚠

Apple’s documentation

kCFErrorDomainOSStatus⚠

Apple’s documentation

kCFErrorDomainPOSIX⚠

Apple’s documentation

kCFErrorFilePathKey⚠

Apple’s documentation

kCFErrorLocalizedDescriptionKey⚠

Apple’s documentation

kCFErrorLocalizedFailureKey⚠

Apple’s documentation

kCFErrorLocalizedFailureReasonKey⚠

Apple’s documentation

kCFErrorLocalizedRecoverySuggestionKey⚠

Apple’s documentation

kCFErrorURLKey⚠

Apple’s documentation

kCFErrorUnderlyingErrorKey⚠

Apple’s documentation

kCFNotFound

Apple’s documentation

kCFNull⚠

Apple’s documentation

kCFNumberNaN⚠

Apple’s documentation

kCFNumberNegativeInfinity⚠

Apple’s documentation

kCFNumberPositiveInfinity⚠

Apple’s documentation

kCFRunLoopCommonModes⚠

Apple’s documentation

kCFRunLoopDefaultMode⚠

Apple’s documentation

kCFStringTransformFullwidthHalfwidth⚠

Apple’s documentation

kCFStringTransformHiraganaKatakana⚠

Apple’s documentation

kCFStringTransformLatinArabic⚠

Apple’s documentation

kCFStringTransformLatinCyrillic⚠

Apple’s documentation

kCFStringTransformLatinGreek⚠

Apple’s documentation

kCFStringTransformLatinHangul⚠

Apple’s documentation

kCFStringTransformLatinHebrew⚠

Apple’s documentation

kCFStringTransformLatinHiragana⚠

Apple’s documentation

kCFStringTransformLatinKatakana⚠

Apple’s documentation

kCFStringTransformLatinThai⚠

Apple’s documentation

kCFStringTransformMandarinLatin⚠

Apple’s documentation

kCFStringTransformStripCombiningMarks⚠

Apple’s documentation

kCFStringTransformStripDiacritics⚠

Apple’s documentation

kCFStringTransformToLatin⚠

Apple’s documentation

kCFStringTransformToUnicodeName⚠

Apple’s documentation

kCFStringTransformToXMLHex⚠

Apple’s documentation

kCFTypeArrayCallBacks⚠

Predefined CFArrayCallBacks structure containing a set of callbacks appropriate for use when the values in a CFArray are all CFTypes.

kCFTypeDictionaryKeyCallBacks⚠

Predefined CFDictionaryKeyCallBacks structure containing a set of callbacks appropriate for use when the keys of a CFDictionary are all CFTypes.

kCFTypeDictionaryValueCallBacks⚠

Predefined CFDictionaryValueCallBacks structure containing a set of callbacks appropriate for use when the values in a CFDictionary are all CFTypes.

kCFURLAddedToDirectoryDateKey⚠

Apple’s documentation

kCFURLApplicationIsScriptableKey⚠

Apple’s documentation

kCFURLAttributeModificationDateKey⚠

Apple’s documentation

kCFURLCanonicalPathKey⚠

Apple’s documentation

kCFURLContentAccessDateKey⚠

Apple’s documentation

kCFURLContentModificationDateKey⚠

Apple’s documentation

kCFURLCreationDateKey⚠

Apple’s documentation

kCFURLCustomIconKey⚠Deprecated

Apple’s documentation

kCFURLDirectoryEntryCountKey⚠

Apple’s documentation

kCFURLDocumentIdentifierKey⚠

Apple’s documentation

kCFURLEffectiveIconKey⚠Deprecated

Apple’s documentation

kCFURLFileAllocatedSizeKey⚠

Apple’s documentation

kCFURLFileContentIdentifierKey⚠

Apple’s documentation

kCFURLFileIdentifierKey⚠

Apple’s documentation

kCFURLFileProtectionComplete⚠

Apple’s documentation

kCFURLFileProtectionCompleteUnlessOpen⚠

Apple’s documentation

kCFURLFileProtectionCompleteUntilFirstUserAuthentication⚠

Apple’s documentation

kCFURLFileProtectionCompleteWhenUserInactive⚠

Apple’s documentation

kCFURLFileProtectionKey⚠

Apple’s documentation

kCFURLFileProtectionNone⚠

Apple’s documentation

kCFURLFileResourceIdentifierKey⚠

Apple’s documentation

kCFURLFileResourceTypeBlockSpecial⚠

Apple’s documentation

kCFURLFileResourceTypeCharacterSpecial⚠

Apple’s documentation

kCFURLFileResourceTypeDirectory⚠

Apple’s documentation

kCFURLFileResourceTypeKey⚠

Apple’s documentation

kCFURLFileResourceTypeNamedPipe⚠

Apple’s documentation

kCFURLFileResourceTypeRegular⚠

Apple’s documentation

kCFURLFileResourceTypeSocket⚠

Apple’s documentation

kCFURLFileResourceTypeSymbolicLink⚠

Apple’s documentation

kCFURLFileResourceTypeUnknown⚠

Apple’s documentation

kCFURLFileSecurityKey⚠

Apple’s documentation

kCFURLFileSizeKey⚠

Apple’s documentation

kCFURLGenerationIdentifierKey⚠

Apple’s documentation

kCFURLHasHiddenExtensionKey⚠

Apple’s documentation

kCFURLIsAliasFileKey⚠

Apple’s documentation

kCFURLIsApplicationKey⚠

Apple’s documentation

kCFURLIsDirectoryKey⚠

Apple’s documentation

kCFURLIsExcludedFromBackupKey⚠

Apple’s documentation

kCFURLIsExecutableKey⚠

Apple’s documentation

kCFURLIsHiddenKey⚠

Apple’s documentation

kCFURLIsMountTriggerKey⚠

Apple’s documentation

kCFURLIsPackageKey⚠

Apple’s documentation

kCFURLIsPurgeableKey⚠

Apple’s documentation

kCFURLIsReadableKey⚠

Apple’s documentation

kCFURLIsRegularFileKey⚠

Apple’s documentation

kCFURLIsSparseKey⚠

Apple’s documentation

kCFURLIsSymbolicLinkKey⚠

Apple’s documentation

kCFURLIsSystemImmutableKey⚠

Apple’s documentation

kCFURLIsUbiquitousItemKey⚠

Apple’s documentation

kCFURLIsUserImmutableKey⚠

Apple’s documentation

kCFURLIsVolumeKey⚠

Apple’s documentation

kCFURLIsWritableKey⚠

Apple’s documentation

kCFURLKeysOfUnsetValuesKey⚠

Apple’s documentation

kCFURLLabelColorKey⚠Deprecated

Apple’s documentation

kCFURLLabelNumberKey⚠

Apple’s documentation

kCFURLLinkCountKey⚠

Apple’s documentation

kCFURLLocalizedLabelKey⚠

Apple’s documentation

kCFURLLocalizedNameKey⚠

Apple’s documentation

kCFURLLocalizedTypeDescriptionKey⚠

Apple’s documentation

kCFURLMayHaveExtendedAttributesKey⚠

Apple’s documentation

kCFURLMayShareFileContentKey⚠

Apple’s documentation

kCFURLNameKey⚠

Apple’s documentation

kCFURLParentDirectoryURLKey⚠

Apple’s documentation

kCFURLPathKey⚠

Apple’s documentation

kCFURLPreferredIOBlockSizeKey⚠

Apple’s documentation

kCFURLQuarantinePropertiesKey⚠

Apple’s documentation

kCFURLTagNamesKey⚠

Apple’s documentation

kCFURLTotalFileAllocatedSizeKey⚠

Apple’s documentation

kCFURLTotalFileSizeKey⚠

Apple’s documentation

kCFURLTypeIdentifierKey⚠Deprecated

Apple’s documentation

kCFURLUbiquitousItemDownloadingErrorKey⚠

Apple’s documentation

kCFURLUbiquitousItemDownloadingStatusCurrent⚠

Apple’s documentation

kCFURLUbiquitousItemDownloadingStatusDownloaded⚠

Apple’s documentation

kCFURLUbiquitousItemDownloadingStatusKey⚠

Apple’s documentation

kCFURLUbiquitousItemDownloadingStatusNotDownloaded⚠

Apple’s documentation

kCFURLUbiquitousItemHasUnresolvedConflictsKey⚠

Apple’s documentation

kCFURLUbiquitousItemIsDownloadedKey⚠Deprecated

Apple’s documentation

kCFURLUbiquitousItemIsDownloadingKey⚠

Apple’s documentation

kCFURLUbiquitousItemIsExcludedFromSyncKey⚠

Apple’s documentation

kCFURLUbiquitousItemIsSyncPausedKey⚠

Apple’s documentation

kCFURLUbiquitousItemIsUploadedKey⚠

Apple’s documentation

kCFURLUbiquitousItemIsUploadingKey⚠

Apple’s documentation

kCFURLUbiquitousItemPercentDownloadedKey⚠Deprecated

Apple’s documentation

kCFURLUbiquitousItemPercentUploadedKey⚠Deprecated

Apple’s documentation

kCFURLUbiquitousItemSupportedSyncControlsKey⚠

Apple’s documentation

kCFURLUbiquitousItemUploadingErrorKey⚠

Apple’s documentation

kCFURLVolumeAvailableCapacityForImportantUsageKey⚠

Apple’s documentation

kCFURLVolumeAvailableCapacityForOpportunisticUsageKey⚠

Apple’s documentation

kCFURLVolumeAvailableCapacityKey⚠

Apple’s documentation

kCFURLVolumeCreationDateKey⚠

Apple’s documentation

kCFURLVolumeIdentifierKey⚠

Apple’s documentation

kCFURLVolumeIsAutomountedKey⚠

Apple’s documentation

kCFURLVolumeIsBrowsableKey⚠

Apple’s documentation

kCFURLVolumeIsEjectableKey⚠

Apple’s documentation

kCFURLVolumeIsEncryptedKey⚠

Apple’s documentation

kCFURLVolumeIsInternalKey⚠

Apple’s documentation

kCFURLVolumeIsJournalingKey⚠

Apple’s documentation

kCFURLVolumeIsLocalKey⚠

Apple’s documentation

kCFURLVolumeIsReadOnlyKey⚠

Apple’s documentation

kCFURLVolumeIsRemovableKey⚠

Apple’s documentation

kCFURLVolumeIsRootFileSystemKey⚠

Apple’s documentation

kCFURLVolumeLocalizedFormatDescriptionKey⚠

Apple’s documentation

kCFURLVolumeLocalizedNameKey⚠

Apple’s documentation

kCFURLVolumeMaximumFileSizeKey⚠

Apple’s documentation

kCFURLVolumeMountFromLocationKey⚠

Apple’s documentation

kCFURLVolumeNameKey⚠

Apple’s documentation

kCFURLVolumeResourceCountKey⚠

Apple’s documentation

kCFURLVolumeSubtypeKey⚠

Apple’s documentation

kCFURLVolumeSupportsAccessPermissionsKey⚠

Apple’s documentation

kCFURLVolumeSupportsAdvisoryFileLockingKey⚠

Apple’s documentation

kCFURLVolumeSupportsCasePreservedNamesKey⚠

Apple’s documentation

kCFURLVolumeSupportsCaseSensitiveNamesKey⚠

Apple’s documentation

kCFURLVolumeSupportsCompressionKey⚠

Apple’s documentation

kCFURLVolumeSupportsExclusiveRenamingKey⚠

Apple’s documentation

kCFURLVolumeSupportsExtendedSecurityKey⚠

Apple’s documentation

kCFURLVolumeSupportsFileCloningKey⚠

Apple’s documentation

kCFURLVolumeSupportsFileProtectionKey⚠

Apple’s documentation

kCFURLVolumeSupportsHardLinksKey⚠

Apple’s documentation

kCFURLVolumeSupportsImmutableFilesKey⚠

Apple’s documentation

kCFURLVolumeSupportsJournalingKey⚠

Apple’s documentation

kCFURLVolumeSupportsPersistentIDsKey⚠

Apple’s documentation

kCFURLVolumeSupportsRenamingKey⚠

Apple’s documentation

kCFURLVolumeSupportsRootDirectoryDatesKey⚠

Apple’s documentation

kCFURLVolumeSupportsSparseFilesKey⚠

Apple’s documentation

kCFURLVolumeSupportsSwapRenamingKey⚠

Apple’s documentation

kCFURLVolumeSupportsSymbolicLinksKey⚠

Apple’s documentation

kCFURLVolumeSupportsVolumeSizesKey⚠

Apple’s documentation

kCFURLVolumeSupportsZeroRunsKey⚠

Apple’s documentation

kCFURLVolumeTotalCapacityKey⚠

Apple’s documentation

kCFURLVolumeTypeNameKey⚠

Apple’s documentation

kCFURLVolumeURLForRemountingKey⚠

Apple’s documentation

kCFURLVolumeURLKey⚠

Apple’s documentation

kCFURLVolumeUUIDStringKey⚠

Apple’s documentation

Traits

ConcreteType

A concrete CoreFoundation type.

Type

A CoreFoundation-like type.

Functions

CFAllocatorAllocateDeprecated

CFAllocatorAllocateBytesDeprecated

CFAllocatorAllocateTyped⚠Deprecated

CFAllocatorCreate⚠Deprecated

CFAllocatorDeallocate⚠Deprecated

CFAllocatorGetContext⚠Deprecated

CFAllocatorGetDefaultDeprecated

CFAllocatorGetPreferredSizeForSizeDeprecated

CFAllocatorReallocate⚠Deprecated

CFAllocatorReallocateBytes⚠Deprecated

CFAllocatorReallocateTyped⚠Deprecated

CFAllocatorSetDefaultDeprecated

CFArrayAppendArray⚠Deprecated

CFArrayAppendValue⚠Deprecated

CFArrayApplyFunction⚠Deprecated

CFArrayBSearchValues⚠Deprecated

CFArrayContainsValue⚠Deprecated

CFArrayCreate⚠Deprecated

CFArrayCreateCopy⚠Deprecated

CFArrayCreateMutable⚠Deprecated

CFArrayCreateMutableCopy⚠Deprecated

CFArrayExchangeValuesAtIndices⚠Deprecated

CFArrayGetCountDeprecated

CFArrayGetCountOfValue⚠Deprecated

CFArrayGetFirstIndexOfValue⚠Deprecated

CFArrayGetLastIndexOfValue⚠Deprecated

CFArrayGetValueAtIndex⚠Deprecated

CFArrayGetValues⚠Deprecated

CFArrayInsertValueAtIndex⚠Deprecated

CFArrayRemoveAllValuesDeprecated

CFArrayRemoveValueAtIndex⚠Deprecated

CFArrayReplaceValues⚠Deprecated

CFArraySetValueAtIndex⚠Deprecated

CFArraySortValues⚠Deprecated

CFBooleanGetValueDeprecated

CFCopyDescription

CFCopyTypeIDDescription

CFDataAppendBytes⚠Deprecated

CFDataCreate⚠Deprecated

CFDataCreateCopyDeprecated

CFDataCreateMutableDeprecated

CFDataCreateMutableCopy⚠Deprecated

CFDataCreateWithBytesNoCopy⚠Deprecated

CFDataDeleteBytesDeprecated

CFDataFind⚠Deprecated

CFDataGetBytePtrDeprecated

CFDataGetBytes⚠Deprecated

CFDataGetLengthDeprecated

CFDataGetMutableBytePtrDeprecated

CFDataIncreaseLengthDeprecated

CFDataReplaceBytes⚠Deprecated

CFDataSetLengthDeprecated

CFDictionaryAddValue⚠Deprecated

CFDictionaryApplyFunction⚠Deprecated

CFDictionaryContainsKey⚠Deprecated

CFDictionaryContainsValue⚠Deprecated

CFDictionaryCreate⚠Deprecated

CFDictionaryCreateCopyDeprecated

CFDictionaryCreateMutable⚠Deprecated

CFDictionaryCreateMutableCopy⚠Deprecated

CFDictionaryGetCountDeprecated

CFDictionaryGetCountOfKey⚠Deprecated

CFDictionaryGetCountOfValue⚠Deprecated

CFDictionaryGetKeysAndValues⚠Deprecated

CFDictionaryGetValue⚠Deprecated

CFDictionaryGetValueIfPresent⚠Deprecated

CFDictionaryRemoveAllValuesDeprecated

CFDictionaryRemoveValue⚠Deprecated

CFDictionaryReplaceValue⚠Deprecated

CFDictionarySetValue⚠Deprecated

CFEqual

CFErrorCopyDescriptionDeprecated

CFErrorCopyFailureReasonDeprecated

CFErrorCopyRecoverySuggestionDeprecated

CFErrorCopyUserInfoDeprecated

CFErrorCreate⚠Deprecated

CFErrorCreateWithUserInfoKeysAndValues⚠Deprecated

CFErrorGetCodeDeprecated

CFErrorGetDomainDeprecated

CFGetAllocator

CFGetRetainCount

CFGetTypeID

CFHash

CFNumberCompare⚠Deprecated

CFNumberCreate⚠Deprecated

CFNumberGetByteSizeDeprecated

CFNumberGetTypeDeprecated

CFNumberGetValue⚠Deprecated

CFNumberIsFloatTypeDeprecated

CFRunLoopAddCommonModeDeprecated

CFRunLoopAddObserverDeprecated

CFRunLoopAddSourceDeprecated

CFRunLoopAddTimerDeprecated

CFRunLoopContainsObserverDeprecated

CFRunLoopContainsSourceDeprecated

CFRunLoopContainsTimerDeprecated

CFRunLoopCopyAllModesDeprecated

CFRunLoopCopyCurrentModeDeprecated

CFRunLoopGetCurrentDeprecated

CFRunLoopGetMainDeprecated

CFRunLoopIsWaitingDeprecated

CFRunLoopObserverCreate⚠Deprecated

CFRunLoopObserverDoesRepeatDeprecated

CFRunLoopObserverGetActivitiesDeprecated

CFRunLoopObserverGetContext⚠Deprecated

CFRunLoopObserverGetOrderDeprecated

CFRunLoopObserverInvalidateDeprecated

CFRunLoopObserverIsValidDeprecated

CFRunLoopRemoveObserverDeprecated

CFRunLoopRemoveSourceDeprecated

CFRunLoopRemoveTimerDeprecated

CFRunLoopRunDeprecated

CFRunLoopSourceCreate⚠Deprecated

CFRunLoopSourceGetContext⚠Deprecated

CFRunLoopSourceGetOrderDeprecated

CFRunLoopSourceInvalidateDeprecated

CFRunLoopSourceIsValidDeprecated

CFRunLoopSourceSignalDeprecated

CFRunLoopStopDeprecated

CFRunLoopTimerDoesRepeatDeprecated

CFRunLoopTimerGetContext⚠Deprecated

CFRunLoopTimerGetOrderDeprecated

CFRunLoopTimerInvalidateDeprecated

CFRunLoopTimerIsValidDeprecated

CFRunLoopWakeUpDeprecated

CFShow

CFShowStr

CFStringAppendDeprecated

CFStringAppendCString⚠Deprecated

CFStringAppendCharacters⚠Deprecated

CFStringAppendPascalString⚠Deprecated

CFStringCompareDeprecated

CFStringCompareWithOptions⚠Deprecated

CFStringConvertEncodingToIANACharSetNameDeprecated

CFStringConvertEncodingToNSStringEncodingDeprecated

CFStringConvertEncodingToWindowsCodepageDeprecated

CFStringConvertIANACharSetNameToEncodingDeprecated

CFStringConvertNSStringEncodingToEncodingDeprecated

CFStringConvertWindowsCodepageToEncodingDeprecated

CFStringCreateArrayBySeparatingStringsDeprecated

CFStringCreateArrayWithFindResults⚠Deprecated

CFStringCreateByCombiningStrings⚠Deprecated

CFStringCreateCopyDeprecated

CFStringCreateExternalRepresentationDeprecated

CFStringCreateFromExternalRepresentationDeprecated

CFStringCreateMutableDeprecated

CFStringCreateMutableCopyDeprecated

CFStringCreateMutableWithExternalCharactersNoCopy⚠Deprecated

CFStringCreateWithBytes⚠Deprecated

CFStringCreateWithBytesNoCopy⚠Deprecated

CFStringCreateWithCString⚠Deprecated

CFStringCreateWithCStringNoCopy⚠Deprecated

CFStringCreateWithCharacters⚠Deprecated

CFStringCreateWithCharactersNoCopy⚠Deprecated

CFStringCreateWithFileSystemRepresentation⚠Deprecated

CFStringCreateWithPascalString⚠Deprecated

CFStringCreateWithPascalStringNoCopy⚠Deprecated

CFStringCreateWithSubstring⚠Deprecated

CFStringDelete⚠Deprecated

CFStringFindDeprecated

CFStringFindAndReplace⚠Deprecated

CFStringFindWithOptions⚠Deprecated

CFStringGetBytes⚠Deprecated

CFStringGetCString⚠Deprecated

CFStringGetCStringPtrDeprecated

CFStringGetCharacterAtIndex⚠Deprecated

CFStringGetCharacters⚠Deprecated

CFStringGetCharactersPtrDeprecated

CFStringGetDoubleValueDeprecated

CFStringGetFastestEncodingDeprecated

CFStringGetFileSystemRepresentation⚠Deprecated

CFStringGetIntValueDeprecated

CFStringGetLengthDeprecated

CFStringGetLineBounds⚠Deprecated

CFStringGetListOfAvailableEncodingsDeprecated

CFStringGetMaximumSizeForEncodingDeprecated

CFStringGetMaximumSizeOfFileSystemRepresentationDeprecated

CFStringGetMostCompatibleMacStringEncodingDeprecated

CFStringGetNameOfEncodingDeprecated

CFStringGetParagraphBounds⚠Deprecated

CFStringGetPascalString⚠Deprecated

CFStringGetPascalStringPtrDeprecated

CFStringGetRangeOfComposedCharactersAtIndex⚠Deprecated

CFStringGetSmallestEncodingDeprecated

CFStringGetSystemEncodingDeprecated

CFStringHasPrefixDeprecated

CFStringHasSuffixDeprecated

CFStringInsert⚠Deprecated

CFStringIsEncodingAvailableDeprecated

CFStringNormalize⚠Deprecated

CFStringPad⚠Deprecated

CFStringReplace⚠Deprecated

CFStringReplaceAllDeprecated

CFStringSetExternalCharactersNoCopy⚠Deprecated

CFStringTransform⚠Deprecated

CFStringTrimDeprecated

CFStringTrimWhitespaceDeprecated

CFURLCanBeDecomposedDeprecated

CFURLClearResourcePropertyCacheDeprecated

CFURLClearResourcePropertyCacheForKeyDeprecated

CFURLCopyAbsoluteURLDeprecated

CFURLCopyFileSystemPathDeprecated

CFURLCopyFragmentDeprecated

CFURLCopyHostNameDeprecated

CFURLCopyLastPathComponentDeprecated

CFURLCopyNetLocationDeprecated

CFURLCopyParameterStringDeprecated

CFURLCopyPasswordDeprecated

CFURLCopyPathDeprecated

CFURLCopyPathExtensionDeprecated

CFURLCopyQueryStringDeprecated

CFURLCopyResourcePropertiesForKeys⚠Deprecated

CFURLCopyResourcePropertyForKey⚠Deprecated

CFURLCopyResourceSpecifierDeprecated

CFURLCopySchemeDeprecated

CFURLCopyStrictPath⚠Deprecated

CFURLCopyUserNameDeprecated

CFURLCreateAbsoluteURLWithBytes⚠Deprecated

CFURLCreateBookmarkData⚠Deprecated

CFURLCreateBookmarkDataFromAliasRecord⚠Deprecated

CFURLCreateBookmarkDataFromFile⚠Deprecated

CFURLCreateByResolvingBookmarkData⚠Deprecated

CFURLCreateCopyAppendingPathComponentDeprecated

CFURLCreateCopyAppendingPathExtensionDeprecated

CFURLCreateCopyDeletingLastPathComponentDeprecated

CFURLCreateCopyDeletingPathExtensionDeprecated

CFURLCreateDataDeprecated

CFURLCreateFilePathURL⚠Deprecated

CFURLCreateFileReferenceURL⚠Deprecated

CFURLCreateFromFileSystemRepresentation⚠Deprecated

CFURLCreateFromFileSystemRepresentationRelativeToBase⚠Deprecated

CFURLCreateResourcePropertiesForKeysFromBookmarkData⚠Deprecated

CFURLCreateResourcePropertyForKeyFromBookmarkData⚠Deprecated

CFURLCreateStringByAddingPercentEscapes⚠Deprecated

CFURLCreateStringByReplacingPercentEscapesDeprecated

CFURLCreateStringByReplacingPercentEscapesUsingEncoding⚠Deprecated

CFURLCreateWithBytes⚠Deprecated

CFURLCreateWithFileSystemPathDeprecated

CFURLCreateWithFileSystemPathRelativeToBaseDeprecated

CFURLCreateWithStringDeprecated

CFURLGetBaseURLDeprecated

CFURLGetByteRangeForComponent⚠Deprecated

CFURLGetBytes⚠Deprecated

CFURLGetFileSystemRepresentation⚠Deprecated

CFURLGetPortNumberDeprecated

CFURLGetStringDeprecated

CFURLHasDirectoryPathDeprecated

CFURLIsFileReferenceURLDeprecated

CFURLResourceIsReachable⚠Deprecated

CFURLSetResourcePropertiesForKeys⚠Deprecated

CFURLSetResourcePropertyForKey⚠Deprecated

CFURLSetTemporaryResourcePropertyForKeyDeprecated

CFURLStartAccessingSecurityScopedResource⚠Deprecated

CFURLStopAccessingSecurityScopedResource⚠Deprecated

CFURLWriteBookmarkDataToFile⚠Deprecated

CFUUIDCreateDeprecated

CFUUIDCreateFromStringDeprecated

CFUUIDCreateFromUUIDBytesDeprecated

CFUUIDCreateStringDeprecated

CFUUIDCreateWithBytesDeprecated

CFUUIDGetConstantUUIDWithBytesDeprecated

CFUUIDGetUUIDBytesDeprecated

Type Aliases

CFAllocatorAllocateCallBack

Apple’s documentation

CFAllocatorCopyDescriptionCallBack

Apple’s documentation

CFAllocatorDeallocateCallBack

Apple’s documentation

CFAllocatorPreferredSizeCallBack

Apple’s documentation

CFAllocatorReallocateCallBack

Apple’s documentation

CFAllocatorReleaseCallBack

Apple’s documentation

CFAllocatorRetainCallBack

Apple’s documentation

CFAllocatorTypeID

Apple’s documentation

CFArrayApplierFunction

Type of the callback function used by the apply functions of CFArrays.

CFArrayCopyDescriptionCallBack

Apple’s documentation

CFArrayEqualCallBack

Apple’s documentation

CFArrayReleaseCallBack

Apple’s documentation

CFArrayRetainCallBack

Structure containing the callbacks of a CFArray. Field: version The version number of the structure type being passed in as a parameter to the CFArray creation functions. This structure is version 0. Field: retain The callback used to add a retain for the array on values as they are put into the array. This callback returns the value to store in the array, which is usually the value parameter passed to this callback, but may be a different value if a different value should be stored in the array. The array’s allocator is passed as the first argument. Field: release The callback used to remove a retain previously added for the array from values as they are removed from the array. The array’s allocator is passed as the first argument. Field: copyDescription The callback used to create a descriptive string representation of each value in the array. This is used by the CFCopyDescription() function. Field: equal The callback used to compare values in the array for equality for some operations.

CFComparatorFunction

Apple’s documentation

CFDictionaryApplierFunction

Type of the callback function used by the apply functions of CFDictionarys.

CFDictionaryCopyDescriptionCallBack

Apple’s documentation

CFDictionaryEqualCallBack

Apple’s documentation

CFDictionaryHashCallBack

Apple’s documentation

CFDictionaryReleaseCallBack

Apple’s documentation

CFDictionaryRetainCallBack

Structure containing the callbacks for keys of a CFDictionary. Field: version The version number of the structure type being passed in as a parameter to the CFDictionary creation functions. This structure is version 0. Field: retain The callback used to add a retain for the dictionary on keys as they are used to put values into the dictionary. This callback returns the value to use as the key in the dictionary, which is usually the value parameter passed to this callback, but may be a different value if a different value should be used as the key. The dictionary’s allocator is passed as the first argument. Field: release The callback used to remove a retain previously added for the dictionary from keys as their values are removed from the dictionary. The dictionary’s allocator is passed as the first argument. Field: copyDescription The callback used to create a descriptive string representation of each key in the dictionary. This is used by the CFCopyDescription() function. Field: equal The callback used to compare keys in the dictionary for equality. Field: hash The callback used to compute a hash code for keys as they are used to access, add, or remove values in the dictionary.

CFErrorDomain

Apple’s documentation

CFHashCode

Apple’s documentation

CFIndex

Apple’s documentation

CFOptionFlags

Apple’s documentation

CFPropertyList

Apple’s documentation

CFRunLoopMode

Apple’s documentation

CFRunLoopObserverCallBack

Apple’s documentation

CFRunLoopTimerCallBack

Apple’s documentation

CFStringEncoding

Apple’s documentation

CFTypeID

Apple’s documentation

CFURLBookmarkFileCreationOptions

Apple’s documentation

HRESULT

Apple’s documentation

LPVOID

Apple’s documentation

REFIID

Apple’s documentation

ULONG

Apple’s documentation