objc2_core_foundation

Function CFArrayAppendArray

pub unsafe extern "C-unwind" fn CFArrayAppendArray(
    the_array: Option<&CFMutableArray>,
    other_array: Option<&CFArray>,
    other_range: CFRange,
)

Adds the values from an array to another array.

Parameter theArray: The array to which values from the otherArray are to be added. If this parameter is not a valid mutable CFArray, the behavior is undefined.

Parameter otherArray: The array providing the values to be added to the array. If this parameter is not a valid CFArray, the behavior is undefined.

Parameter otherRange: The range within the otherArray from which to add the values to the array. If the range location or end point (defined by the location plus length minus 1) is outside the index space of the otherArray (0 to N-1 inclusive, where N is the count of the otherArray), the behavior is undefined. The new values are retained by the array using the retain callback provided when the array was created. If the values are not of the sort expected by the retain callback, the behavior is undefined. The values are assigned to the indices one larger than the previous largest index in the array, and beyond, and the count of the array is increased by range.length. The values are assigned new indices in the array from smallest to largest index in the order in which they appear in the otherArray.