objc2_core_foundation

Function CFArrayBSearchValues

pub unsafe extern "C-unwind" fn CFArrayBSearchValues(
    the_array: &CFArray,
    range: CFRange,
    value: *const c_void,
    comparator: CFComparatorFunction,
    context: *mut c_void,
) -> CFIndex

Searches the array for the value using a binary search algorithm.

Parameter theArray: The array to be searched. If this parameter is not a valid CFArray, the behavior is undefined. If the array is not sorted from least to greatest according to the comparator function, the behavior is undefined.

Parameter range: The range within the array to search. If the range location or end point (defined by the location plus length minus 1) is outside the index space of the array (0 to N-1 inclusive, where N is the count of the array), the behavior is undefined. If the range length is negative, the behavior is undefined. The range may be empty (length 0).

Parameter value: The value for which to find a match in the array. If value, or any of the values in the array, are not understood by the comparator callback, the behavior is undefined.

Parameter comparator: The function with the comparator function type signature which is used in the binary search operation to compare values in the array with the given value. If this parameter is not a pointer to a function of the correct prototype, the behavior is undefined. If there are values in the range which the comparator function does not expect or cannot properly compare, the behavior is undefined.

Parameter context: A pointer-sized user-defined value, which is passed as the third parameter to the comparator function, but is otherwise unused by this function. If the context is not what is expected by the comparator function, the behavior is undefined.

Returns: The return value is either 1) the index of a value that matched, if the target value matches one or more in the range, 2) greater than or equal to the end point of the range, if the value is greater than all the values in the range, or 3) the index of the value greater than the target value, if the value lies between two of (or less than all of) the values in the range.