objc2_core_foundation

Function CFDictionaryCreateMutableCopy

pub unsafe extern "C-unwind" fn CFDictionaryCreateMutableCopy(
    allocator: Option<&CFAllocator>,
    capacity: CFIndex,
    the_dict: Option<&CFDictionary>,
) -> Option<CFRetained<CFMutableDictionary>>

Creates a new mutable dictionary with the key-value pairs from the given dictionary.

Parameter allocator: The CFAllocator which should be used to allocate memory for the dictionary and its storage for values. This parameter may be NULL in which case the current default CFAllocator is used. If this reference is not a valid CFAllocator, the behavior is undefined.

Parameter capacity: A hint about the number of values that will be held by the CFDictionary. Pass 0 for no hint. The implementation may ignore this hint, or may use it to optimize various operations. A dictionary’s actual capacity is only limited by address space and available memory constraints). This parameter must be greater than or equal to the count of the dictionary which is to be copied, or the behavior is undefined. If this parameter is negative, the behavior is undefined.

Parameter theDict: The dictionary which is to be copied. The keys and values from the dictionary are copied as pointers into the new dictionary (that is, the values themselves are copied, not that which the values point to, if anything). However, the keys and values are also retained by the new dictionary using the retain function of the original dictionary. The count of the new dictionary will be the same as the given dictionary. The new dictionary uses the same callbacks as the dictionary to be copied. If this parameter is not a valid CFDictionary, the behavior is undefined.

Returns: A reference to the new mutable CFDictionary.