objc2_core_foundation

Function CFDictionaryGetKeysAndValues

pub unsafe extern "C-unwind" fn CFDictionaryGetKeysAndValues(
    the_dict: &CFDictionary,
    keys: *mut *const c_void,
    values: *mut *const c_void,
)

Fills the two buffers with the keys and values from the dictionary.

Parameter theDict: The dictionary to be queried. If this parameter is not a valid CFDictionary, the behavior is undefined.

Parameter keys: A C array of pointer-sized values to be filled with keys from the dictionary. The keys and values C arrays are parallel to each other (that is, the items at the same indices form a key-value pair from the dictionary). This parameter may be NULL if the keys are not desired. If this parameter is not a valid pointer to a C array of at least CFDictionaryGetCount() pointers, or NULL, the behavior is undefined.

Parameter values: A C array of pointer-sized values to be filled with values from the dictionary. The keys and values C arrays are parallel to each other (that is, the items at the same indices form a key-value pair from the dictionary). This parameter may be NULL if the values are not desired. If this parameter is not a valid pointer to a C array of at least CFDictionaryGetCount() pointers, or NULL, the behavior is undefined.