objc2_core_foundation

Function CFDictionaryGetValue

pub unsafe extern "C-unwind" fn CFDictionaryGetValue(
    the_dict: &CFDictionary,
    key: *const c_void,
) -> *const c_void

Retrieves the value associated with the given key.

Parameter theDict: The dictionary to be queried. If this parameter is not a valid CFDictionary, the behavior is undefined.

Parameter key: The key for which to find a match in the dictionary. The hash() and equal() key callbacks provided when the dictionary was created are used to compare. If the hash() key callback was NULL, the key is treated as a pointer and converted to an integer. If the equal() key callback was NULL, pointer equality (in C, ==) is used. If key, or any of the keys in the dictionary, are not understood by the equal() callback, the behavior is undefined.

Returns: The value with the given key in the dictionary, or NULL if no key-value pair with a matching key exists. Since NULL can be a valid value in some dictionaries, the function CFDictionaryGetValueIfPresent() must be used to distinguish NULL-no-found from NULL-is-the-value.