objc2_core_foundation

Function CFErrorCopyDescription

pub unsafe extern "C-unwind" fn CFErrorCopyDescription(
    err: &CFError,
) -> Option<CFRetained<CFString>>

Returns a human-presentable description for the error. CFError creators should strive to make sure the return value is human-presentable and localized by providing a value for kCFErrorLocalizedDescriptionKey at the time of CFError creation.

This is a complete sentence or two which says what failed and why it failed. Please refer to header comments for -[NSError localizedDescription] for details on the steps used to compute this; but roughly:

Use value of kCFErrorLocalizedDescriptionKey as-is if provided.

Use value of kCFErrorLocalizedFailureKey if provided, optionally followed by kCFErrorLocalizedFailureReasonKey if available.

Use value of kCFErrorLocalizedFailureReasonKey, combining with a generic failure message such as: “Operation code not be completed. “ + kCFErrorLocalizedFailureReasonKey.

If all of the above fail, generate a semi-user presentable string from kCFErrorDescriptionKey, the domain, and code. Something like: “Operation could not be completed. Error domain/code occurred. “ or “Operation could not be completed. “ + kCFErrorDescriptionKey + “ (Error domain/code)” Toll-free bridged NSError instances might provide additional behaviors for manufacturing a description string. Do not count on the exact contents or format of the returned string, it might change.

Parameter err: The CFError whose description is to be returned. If this reference is not a valid CFError, the behavior is undefined.

Returns: A CFString with human-presentable description of the CFError. Never NULL.