objc2_core_foundation

Function CFErrorCopyFailureReason

pub unsafe extern "C-unwind" fn CFErrorCopyFailureReason(
    err: &CFError,
) -> Option<CFRetained<CFString>>

Returns a human-presentable failure reason for the error. May return NULL. CFError creators should strive to make sure the return value is human-presentable and localized by providing a value for kCFErrorLocalizedFailureReasonKey at the time of CFError creation.

This is a complete sentence which describes why the operation failed. In many cases this will be just the “because” part of the description (but as a complete sentence, which makes localization easier). By default this looks for kCFErrorLocalizedFailureReasonKey in the user info. Toll-free bridged NSError instances might provide additional behaviors for manufacturing this value. If no user-presentable string is available, returns NULL. Example Description: “Could not save file ‘Letter’ in folder ‘Documents’ because the volume ‘MyDisk’ doesn’t have enough space.” Corresponding FailureReason: “The volume ‘MyDisk’ doesn’t have enough space.”

Parameter err: The CFError whose failure reason is to be returned. If this reference is not a valid CFError, the behavior is undefined.

Returns: A CFString with the localized, end-user presentable failure reason of the CFError, or NULL.