objc2_core_foundation

Function CFSetApplyFunction

pub unsafe extern "C-unwind" fn CFSetApplyFunction(
    the_set: &CFSet,
    applier: CFSetApplierFunction,
    context: *mut c_void,
)

Calls a function once for each value in the set.

Parameter theSet: The set to be operated upon. If this parameter is not a valid CFSet, the behavior is undefined.

Parameter applier: The callback function to call once for each value in the given set. If this parameter is not a pointer to a function of the correct prototype, the behavior is undefined. If there are values in the set which the applier function does not expect or cannot properly apply to, the behavior is undefined.

Parameter context: A pointer-sized user-defined value, which is passed as the second parameter to the applier function, but is otherwise unused by this function. If the context is not what is expected by the applier function, the behavior is undefined.