objc2_core_foundation

Function CFStringTokenizerCopyCurrentTokenAttribute

pub unsafe extern "C-unwind" fn CFStringTokenizerCopyCurrentTokenAttribute(
    tokenizer: &CFStringTokenizer,
    attribute: CFOptionFlags,
) -> Option<CFRetained<CFType>>

Copies the specified attribute of current token.

Parameter tokenizer: The reference to CFStringTokenizer returned by CFStringTokenizerCreate.

Parameter attribute: Specify a token attribute you want to obtain. The value is one of kCFStringTokenizerAttributeLatinTranscription or kCFStringTokenizerAttributeLanguage.

Returns: Token attribute, or NULL if current token does not have the specified attribute or if there is no current token.