objc2_core_foundation

Function CFTreeInsertSibling

pub unsafe extern "C-unwind" fn CFTreeInsertSibling(
    tree: &CFTree,
    new_sibling: Option<&CFTree>,
)

Inserts newSibling into the the parent tree’s linked list of children after tree. The newSibling will have the same parent as tree.

Parameter tree: The tree to insert newSibling after. If this parameter is not a valid CFTree, the behavior is undefined. If the tree does not have a parent, the behavior is undefined.

Parameter newSibling: The sibling to be added. If this parameter is not a valid CFTree, the behavior is undefined. If this parameter is a tree which is already a child of any tree, the behavior is undefined.