objc2_core_foundation

Struct CFAttributedString

pub struct CFAttributedString { /* private fields */ }

Apple’s documentation

This is toll-free bridged with NSAttributedString.

Implementations

impl CFAttributedString

pub unsafe fn new( alloc: Option<&CFAllocator>, str: Option<&CFString>, attributes: Option<&CFDictionary>, ) -> Option<CFRetained<CFAttributedString>>

Creates an attributed string with the specified string and attributes (both copied).

Safety

alloc might not allow None.

str might not allow None.

attributes generic must be of the correct type.

attributes generic must be of the correct type.

attributes might not allow None.

pub unsafe fn with_substring( alloc: Option<&CFAllocator>, a_str: Option<&CFAttributedString>, range: CFRange, ) -> Option<CFRetained<CFAttributedString>>

Creates a sub-attributed string from the specified range. It’s a programming error for range to specify characters outside the bounds of aStr.

Safety

alloc might not allow None.

a_str might not allow None.

pub fn new_copy( alloc: Option<&CFAllocator>, a_str: Option<&CFAttributedString>, ) -> Option<CFRetained<CFAttributedString>>

Creates an immutable attributed string copy.

pub fn string(&self) -> Option<CFRetained<CFString>>

Returns the string for the attributed string. For performance reasons, this will often point at the backing store of the attributed string, and it might change if the attributed string is edited. However, this is an implementation detail, and definitely not something that should be counted on.

pub fn length(&self) -> CFIndex

Returns the length of the attributed string in characters; same as CFStringGetLength(CFAttributedStringGetString(aStr))

pub unsafe fn attributes( &self, loc: CFIndex, effective_range: *mut CFRange, ) -> Option<CFRetained<CFDictionary>>

Returns the attributes at the specified location. If effectiveRange is not NULL, upon return *effectiveRange contains a range over which the exact same set of attributes apply. Note that for performance reasons, the returned effectiveRange is not necessarily the maximal range - for that, use CFAttributedStringGetAttributesAndLongestEffectiveRange(). It’s a programming error for loc to specify a location outside the bounds of the attributed string.

Note that the returned attribute dictionary might change in unpredictable ways from under the caller if the attributed string is edited after this call. If you wish to hang on to the dictionary long-term, you should make an actual copy of it rather than just retaining it. Also, no assumptions should be made about the relationship of the actual CFDictionaryRef returned by this call and the dictionary originally used to set the attributes, other than the fact that the values stored in the dictionary will be identical (that is, ==) to those originally specified.

Safety

effective_range must be a valid pointer.

pub unsafe fn attribute( &self, loc: CFIndex, attr_name: Option<&CFString>, effective_range: *mut CFRange, ) -> Option<CFRetained<CFType>>

Returns the value of a single attribute at the specified location. If the specified attribute doesn’t exist at the location, returns NULL. If effectiveRange is not NULL, upon return *effectiveRange contains a range over which the exact same attribute value applies. Note that for performance reasons, the returned effectiveRange is not necessarily the maximal range - for that, use CFAttributedStringGetAttributeAndLongestEffectiveRange(). It’s a programming error for loc to specify a location outside the bounds of the attributed string.

Safety

attr_name might not allow None.

effective_range must be a valid pointer.

pub unsafe fn attributes_and_longest_effective_range( &self, loc: CFIndex, in_range: CFRange, longest_effective_range: *mut CFRange, ) -> Option<CFRetained<CFDictionary>>

Returns the attributes at the specified location. If longestEffectiveRange is not NULL, upon return *longestEffectiveRange contains the maximal range within inRange over which the exact same set of attributes apply. The returned range is clipped to inRange. It’s a programming error for loc or inRange to specify locations outside the bounds of the attributed string.

Safety

longest_effective_range must be a valid pointer.

pub unsafe fn attribute_and_longest_effective_range( &self, loc: CFIndex, attr_name: Option<&CFString>, in_range: CFRange, longest_effective_range: *mut CFRange, ) -> Option<CFRetained<CFType>>

Returns the value of a single attribute at the specified location. If longestEffectiveRange is not NULL, upon return *longestEffectiveRange contains the maximal range within inRange over which the exact same attribute value applies. The returned range is clipped to inRange. It’s a programming error for loc or inRange to specify locations outside the bounds of the attributed string.

Safety

attr_name might not allow None.

longest_effective_range must be a valid pointer.

impl CFAttributedString

pub unsafe fn bidi_levels_and_resolved_directions( &self, range: CFRange, base_direction: i8, bidi_levels: *mut u8, base_directions: *mut u8, ) -> bool

Fills bidiLevels by applying the Unicode Bidi Algorithm (P, X, W, N, and I) to the characters in range. Returns true if the result is not uni-level LTR (in other words, needing further Bidi processing). baseDirection is NSWritingDirection (NSWritingDirectionNatural, NSWritingDirectionLeftToRight, and NSWritingDirectionRightToLeft). Understands NSWritingDirectionAttributeName values.

Safety

bidi_levels must be a valid pointer.

base_directions must be a valid pointer.

pub unsafe fn statistical_writing_directions( &self, range: CFRange, base_direction: i8, bidi_levels: *mut u8, base_directions: *mut u8, ) -> bool

If baseDirection is not NSWritingDirectionNatural, result comes from CFAttributedStringGetBidiLevelsAndResolvedDirections; otherwise, it fills bidiLevels by applying a statistical approach (a paragraph is RTL if 40% or more of its words are RTL) to the characters in range. Returns true if the result is not uni-level LTR (in other words, needing further Bidi processing). baseDirection is NSWritingDirection (NSWritingDirectionNatural, NSWritingDirectionLeftToRight, and NSWritingDirectionRightToLeft). Understands NSWritingDirectionAttributeName values.

Safety

bidi_levels must be a valid pointer.

base_directions must be a valid pointer.