objc2_core_foundation

Struct CFBundle

pub struct CFBundle { /* private fields */ }

Apple’s documentation

Implementations

impl CFBundle

pub fn main_bundle() -> Option<CFRetained<CFBundle>>

pub fn bundle_with_identifier( bundle_id: Option<&CFString>, ) -> Option<CFRetained<CFBundle>>

pub unsafe fn all_bundles() -> Option<CFRetained<CFArray>>

impl CFBundle

pub fn new( allocator: Option<&CFAllocator>, bundle_url: Option<&CFURL>, ) -> Option<CFRetained<CFBundle>>

pub fn new_bundles_from_directory( allocator: Option<&CFAllocator>, directory_url: Option<&CFURL>, bundle_type: Option<&CFString>, ) -> Option<CFRetained<CFArray>>

pub fn bundle_url(&self) -> Option<CFRetained<CFURL>>

pub fn value_for_info_dictionary_key( &self, key: Option<&CFString>, ) -> Option<CFRetained<CFType>>

pub fn info_dictionary(&self) -> Option<CFRetained<CFDictionary>>

pub fn local_info_dictionary(&self) -> Option<CFRetained<CFDictionary>>

pub unsafe fn package_info( &self, package_type: *mut u32, package_creator: *mut u32, )

Safety

package_type must be a valid pointer.

package_creator must be a valid pointer.

pub fn identifier(&self) -> Option<CFRetained<CFString>>

pub fn version_number(&self) -> u32

pub fn development_region(&self) -> Option<CFRetained<CFString>>

pub fn support_files_directory_url(&self) -> Option<CFRetained<CFURL>>

pub fn resources_directory_url(&self) -> Option<CFRetained<CFURL>>

pub fn private_frameworks_url(&self) -> Option<CFRetained<CFURL>>

pub fn shared_frameworks_url(&self) -> Option<CFRetained<CFURL>>

pub fn shared_support_url(&self) -> Option<CFRetained<CFURL>>

pub fn built_in_plug_ins_url(&self) -> Option<CFRetained<CFURL>>

pub fn info_dictionary_in_directory( bundle_url: Option<&CFURL>, ) -> Option<CFRetained<CFDictionary>>

pub unsafe fn package_info_in_directory( url: Option<&CFURL>, package_type: *mut u32, package_creator: *mut u32, ) -> bool

Safety

url might not allow None.

package_type must be a valid pointer.

package_creator must be a valid pointer.

pub fn resource_url( &self, resource_name: Option<&CFString>, resource_type: Option<&CFString>, sub_dir_name: Option<&CFString>, ) -> Option<CFRetained<CFURL>>

pub fn resource_urls_of_type( &self, resource_type: Option<&CFString>, sub_dir_name: Option<&CFString>, ) -> Option<CFRetained<CFArray>>

pub fn localized_string( &self, key: Option<&CFString>, value: Option<&CFString>, table_name: Option<&CFString>, ) -> Option<CFRetained<CFString>>

pub unsafe fn localized_string_for_localizations( &self, key: Option<&CFString>, value: Option<&CFString>, table_name: Option<&CFString>, localizations: Option<&CFArray>, ) -> Option<CFRetained<CFString>>

Returns a localized string given a list of possible localizations. The one most suitable to use with the given bundle is returned.

Parameters:

bundle: The bundle to examine.

key: The key for the localized string to retrieve.

value: A default value to return if no value exists for key.

tableName: The name of the strings file to search.

localizations: An array of BCP 47 language codes corresponding to available localizations. Bundle compares the array against its available localizations, and uses the best result to retrieve the localized string. If empty, we treat it as no localization is available, and may return a fallback.

Returns: A localized version of the string designated by key in table tableName.

Safety

key might not allow None.

value might not allow None.

table_name might not allow None.

localizations generic must be of the correct type.

localizations might not allow None.

pub fn resource_url_in_directory( bundle_url: Option<&CFURL>, resource_name: Option<&CFString>, resource_type: Option<&CFString>, sub_dir_name: Option<&CFString>, ) -> Option<CFRetained<CFURL>>

pub fn resource_urls_of_type_in_directory( bundle_url: Option<&CFURL>, resource_type: Option<&CFString>, sub_dir_name: Option<&CFString>, ) -> Option<CFRetained<CFArray>>

pub fn bundle_localizations(&self) -> Option<CFRetained<CFArray>>

pub unsafe fn preferred_localizations_from_array( loc_array: Option<&CFArray>, ) -> Option<CFRetained<CFArray>>

Safety

loc_array generic must be of the correct type.

loc_array might not allow None.

pub unsafe fn localizations_for_preferences( loc_array: Option<&CFArray>, pref_array: Option<&CFArray>, ) -> Option<CFRetained<CFArray>>

Safety

loc_array generic must be of the correct type.

loc_array might not allow None.

pref_array generic must be of the correct type.

pref_array might not allow None.

pub fn resource_url_for_localization( &self, resource_name: Option<&CFString>, resource_type: Option<&CFString>, sub_dir_name: Option<&CFString>, localization_name: Option<&CFString>, ) -> Option<CFRetained<CFURL>>

pub fn resource_urls_of_type_for_localization( &self, resource_type: Option<&CFString>, sub_dir_name: Option<&CFString>, localization_name: Option<&CFString>, ) -> Option<CFRetained<CFArray>>

pub fn info_dictionary_for_url( url: Option<&CFURL>, ) -> Option<CFRetained<CFDictionary>>

pub fn localizations_for_url(url: Option<&CFURL>) -> Option<CFRetained<CFArray>>

pub fn executable_architectures_for_url( url: Option<&CFURL>, ) -> Option<CFRetained<CFArray>>

pub fn executable_url(&self) -> Option<CFRetained<CFURL>>

impl CFBundle

pub fn executable_architectures(&self) -> Option<CFRetained<CFArray>>

pub unsafe fn preflight_executable(&self, error: *mut *mut CFError) -> bool

Safety

error must be a valid pointer.

pub unsafe fn load_executable_and_return_error( &self, error: *mut *mut CFError, ) -> bool

Safety

error must be a valid pointer.

pub unsafe fn load_executable(&self) -> bool

pub fn is_executable_loaded(&self) -> bool

pub unsafe fn unload_executable(&self)

pub fn function_pointer_for_name( &self, function_name: Option<&CFString>, ) -> *mut c_void

pub unsafe fn function_pointers_for_names( &self, function_names: Option<&CFArray>, ftbl: *mut *mut c_void, )

Safety

function_names generic must be of the correct type.

function_names might not allow None.

ftbl must be a valid pointer.

pub fn data_pointer_for_name( &self, symbol_name: Option<&CFString>, ) -> *mut c_void

pub unsafe fn data_pointers_for_names( &self, symbol_names: Option<&CFArray>, stbl: *mut *mut c_void, )

Safety

symbol_names generic must be of the correct type.

symbol_names might not allow None.

stbl must be a valid pointer.

pub fn auxiliary_executable_url( &self, executable_name: Option<&CFString>, ) -> Option<CFRetained<CFURL>>

pub fn is_executable_loadable(&self) -> bool

pub fn is_executable_loadable_for_url(url: Option<&CFURL>) -> bool

pub fn is_architecture_loadable(arch: cpu_type_t) -> bool

pub fn plug_in(&self) -> Option<CFRetained<CFPlugIn>>

pub fn open_bundle_resource_map(&self) -> CFBundleRefNum

👎Deprecated: The Carbon Resource Manager is deprecated. This should only be used to access Resource Manager-style resources in old bundles.

pub unsafe fn open_bundle_resource_files( &self, ref_num: *mut CFBundleRefNum, localized_ref_num: *mut CFBundleRefNum, ) -> i32

👎Deprecated: The Carbon Resource Manager is deprecated. This should only be used to access Resource Manager-style resources in old bundles.

Safety

ref_num must be a valid pointer.

localized_ref_num must be a valid pointer.

pub fn close_bundle_resource_map(&self, ref_num: CFBundleRefNum)

👎Deprecated: The Carbon Resource Manager is deprecated. This should only be used to access Resource Manager-style resources in old bundles.