objc2_core_foundation

Struct CFMutableAttributedString

pub struct CFMutableAttributedString { /* private fields */ }

Apple’s documentation

This is toll-free bridged with NSMutableAttributedString.

Implementations

impl CFMutableAttributedString

pub fn new_copy( alloc: Option<&CFAllocator>, max_length: CFIndex, a_str: Option<&CFAttributedString>, ) -> Option<CFRetained<CFMutableAttributedString>>

Creates a mutable attributed string copy. maxLength, if not 0, is a hard bound on the length of the attributed string; exceeding this size limit during any editing operation is a programming error. If 0, there is no limit on the length.

pub fn new( alloc: Option<&CFAllocator>, max_length: CFIndex, ) -> Option<CFRetained<CFMutableAttributedString>>

Creates a mutable empty attributed string. maxLength, if not 0, is a hard bound on the length of the attributed string; exceeding this size limit during any editing operation is a programming error. If 0, there is no limit on the length.

pub unsafe fn replace_string( a_str: Option<&CFMutableAttributedString>, range: CFRange, replacement: Option<&CFString>, )

Modifies the string for the attributed string, much like CFStringReplace(). It’s an error for range to specify characters outside the bounds of aStr.

(Note: This function is a convenience on CFAttributedStringGetMutableString(); however, until CFAttributedStringGetMutableString() is implemented, it remains the only way to edit the string of the attributed string.)

Safety

a_str might not allow None.

replacement might not allow None.

pub fn mutable_string( a_str: Option<&CFMutableAttributedString>, ) -> Option<CFRetained<CFMutableString>>

Gets the string for the attributed string as a mutable string, allowing editing the character contents of the string as if it were an CFMutableString. Attributes corresponding to the edited range are appropriately modified. If, as a result of the edit, new characters are introduced into the string, they inherit the attributes of the first replaced character from range. If no existing characters are replaced by the edit, the new characters inherit the attributes of the character preceding range if it has any, otherwise of the character following range. If the initial string is empty, the attributes for the new characters are also empty.

(Note: This function is not yet implemented and will return NULL except for toll-free bridged instances.)

pub unsafe fn set_attributes( a_str: Option<&CFMutableAttributedString>, range: CFRange, replacement: Option<&CFDictionary>, clear_other_attributes: bool, )

Sets the value of multiple attributes over the specified range, which should be valid. If clearOtherAttributes is false, existing attributes (which aren’t being replaced) are left alone; otherwise they are cleared. The dictionary should be setup for “usual” CF type usage — CFString keys, and arbitrary CFType values. Note that after this call, further mutations to the replacement dictionary argument by the caller will not affect the contents of the attributed string.

Safety

a_str might not allow None.

replacement generic must be of the correct type.

replacement generic must be of the correct type.

replacement might not allow None.

pub unsafe fn set_attribute( a_str: Option<&CFMutableAttributedString>, range: CFRange, attr_name: Option<&CFString>, value: Option<&CFType>, )

Sets the value of a single attribute over the specified range, which should be valid. value should not be NULL.

Safety

a_str might not allow None.

attr_name might not allow None.

value should be of the correct type.

value might not allow None.

pub unsafe fn remove_attribute( a_str: Option<&CFMutableAttributedString>, range: CFRange, attr_name: Option<&CFString>, )

Removes the value of a single attribute over the specified range, which should be valid. It’s OK for the attribute not the exist over the specified range.

Safety

a_str might not allow None.

attr_name might not allow None.

pub unsafe fn replace_attributed_string( a_str: Option<&CFMutableAttributedString>, range: CFRange, replacement: Option<&CFAttributedString>, )

Replaces the attributed substring over the specified range with the attributed string specified in replacement. range should be valid. To delete a range of the attributed string, call CFAttributedStringReplaceString() with empty string and specified range.

Safety

a_str might not allow None.

replacement might not allow None.

pub fn begin_editing(a_str: Option<&CFMutableAttributedString>)

In cases where attributed string might do a bunch of work to assure self-consistency, CFAttributedStringBeginEditing/CFAttributedStringEndEditing allow disabling that to allow deferring and coalescing any work. It’s a good idea to call these around a set of related mutation calls which don’t require the string to be in consistent state in between. These calls can be nested.

pub fn end_editing(a_str: Option<&CFMutableAttributedString>)

In cases where attributed string might do a bunch of work to assure self-consistency, CFAttributedStringBeginEditing/CFAttributedStringEndEditing allow disabling that to allow deferring and coalescing any work. It’s a good idea to call these around a set of related mutation calls which don’t require the string to be in consistent state in between. These calls can be nested.