objc2_core_foundation

Struct CFMutableString

pub struct CFMutableString { /* private fields */ }

Apple’s documentation

This is toll-free bridged with NSMutableString.

Implementations

impl CFMutableString

pub fn new( alloc: Option<&CFAllocator>, max_length: CFIndex, ) -> Option<CFRetained<CFMutableString>>

pub fn new_copy( alloc: Option<&CFAllocator>, max_length: CFIndex, the_string: Option<&CFString>, ) -> Option<CFRetained<CFMutableString>>

pub unsafe fn with_external_characters_no_copy( alloc: Option<&CFAllocator>, chars: *mut u16, num_chars: CFIndex, capacity: CFIndex, external_characters_allocator: Option<&CFAllocator>, ) -> Option<CFRetained<CFMutableString>>

Safety

alloc might not allow None.

chars must be a valid pointer.

external_characters_allocator might not allow None.

impl CFMutableString

pub fn append( the_string: Option<&CFMutableString>, appended_string: Option<&CFString>, )

MutableString functions **

pub unsafe fn append_characters( the_string: Option<&CFMutableString>, chars: *const u16, num_chars: CFIndex, )

Safety

the_string might not allow None.

chars must be a valid pointer.

pub unsafe fn append_pascal_string( the_string: Option<&CFMutableString>, p_str: *const c_char, encoding: CFStringEncoding, )

Safety

the_string might not allow None.

p_str must be a valid pointer.

pub unsafe fn append_c_string( the_string: Option<&CFMutableString>, c_str: *const c_char, encoding: CFStringEncoding, )

Safety

the_string might not allow None.

c_str must be a valid pointer.

pub unsafe fn insert( str: Option<&CFMutableString>, idx: CFIndex, inserted_str: Option<&CFString>, )

Safety

str might not allow None.

inserted_str might not allow None.

pub unsafe fn delete(the_string: Option<&CFMutableString>, range: CFRange)

Safety

the_string might not allow None.

pub unsafe fn replace( the_string: Option<&CFMutableString>, range: CFRange, replacement: Option<&CFString>, )

Safety

the_string might not allow None.

replacement might not allow None.

pub fn replace_all( the_string: Option<&CFMutableString>, replacement: Option<&CFString>, )

pub unsafe fn find_and_replace( the_string: Option<&CFMutableString>, string_to_find: Option<&CFString>, replacement_string: Option<&CFString>, range_to_search: CFRange, compare_options: CFStringCompareFlags, ) -> CFIndex

Safety

the_string might not allow None.

string_to_find might not allow None.

replacement_string might not allow None.

pub unsafe fn set_external_characters_no_copy( the_string: Option<&CFMutableString>, chars: *mut u16, length: CFIndex, capacity: CFIndex, )

Safety

the_string might not allow None.

chars must be a valid pointer.

pub unsafe fn pad( the_string: Option<&CFMutableString>, pad_string: Option<&CFString>, length: CFIndex, index_into_pad: CFIndex, )

Safety

the_string might not allow None.

pad_string might not allow None.

pub fn trim( the_string: Option<&CFMutableString>, trim_string: Option<&CFString>, )

pub fn trim_whitespace(the_string: Option<&CFMutableString>)

pub fn lowercase( the_string: Option<&CFMutableString>, locale: Option<&CFLocale>, )

pub fn uppercase( the_string: Option<&CFMutableString>, locale: Option<&CFLocale>, )

pub fn capitalize( the_string: Option<&CFMutableString>, locale: Option<&CFLocale>, )

impl CFMutableString

pub unsafe fn normalize( the_string: Option<&CFMutableString>, the_form: CFStringNormalizationForm, )

Normalizes the string into the specified form as described in Unicode Technical Report #15.

Parameter theString: The string which is to be normalized. If this parameter is not a valid mutable CFString, the behavior is undefined.

Parameter theForm: The form into which the string is to be normalized. If this parameter is not a valid CFStringNormalizationForm value, the behavior is undefined.

Safety

the_string might not allow None.

pub fn fold( the_string: Option<&CFMutableString>, the_flags: CFStringCompareFlags, the_locale: Option<&CFLocale>, )

Folds the string into the form specified by the flags. Character foldings are operations that convert any of a set of characters sharing similar semantics into a single representative from that set. This function can be used to preprocess strings that are to be compared, searched, or indexed. Note that folding does not include normalization, so it is necessary to use CFStringNormalize in addition to CFStringFold in order to obtain the effect of kCFCompareNonliteral.

Parameter theString: The string which is to be folded. If this parameter is not a valid mutable CFString, the behavior is undefined.

Parameter theFlags: The equivalency flags which describes the character folding form. Only those flags containing the word “insensitive” are recognized here; other flags are ignored. Folding with kCFCompareCaseInsensitive removes case distinctions in accordance with the mapping specified by ftp://ftp.unicode.org/Public/UNIDATA/CaseFolding.txt. Folding with kCFCompareDiacriticInsensitive removes distinctions of accents and other diacritics. Folding with kCFCompareWidthInsensitive removes character width distinctions by mapping characters in the range U+FF00-U+FFEF to their ordinary equivalents.

Parameter theLocale: The locale tailoring the character folding behavior. If NULL, it’s considered to be the system locale returned from CFLocaleGetSystem(). If non-NULL and not a valid CFLocale object, the behavior is undefined.

pub unsafe fn transform( string: Option<&CFMutableString>, range: *mut CFRange, transform: Option<&CFString>, reverse: bool, ) -> bool

Safety

string might not allow None.

range must be a valid pointer.

transform might not allow None.