objc2_core_foundation

Struct CFRetained

pub struct CFRetained<T: ?Sized> { /* private fields */ }

A reference counted pointer type for CoreFoundation types.

CFRetained strongly references or โretainsโ the given object T, and decrements the retain count or โreleasesโ it again when dropped, thereby ensuring it will be deallocated at the right time.

The type T inside CFRetained<T> can be anything that implements Type, i.e. any CoreFoundation-like type.

Comparison to other types

CFRetained<T> is equivalent to objc2::rc::Retained, and can be converted to/from that when the "objc2" feature is enabled. Note though that this uses the underlying CoreFoundation primitives CFRetain / CFRelease / CFAutorelease instead of objc_retain / objc_release / objc_autorelease, to avoid depending on the Objective-C runtime if not needed.

You can also view CFRetained<T> as the CoreFoundation equivalent of std::sync::Arc, that is, it is a thread-safe reference-counting smart pointer that allows cloning by bumping the reference count.

Unlike Arc, objects can be retained directly from a &T using Type::retain (for Arc you need &Arc<T>).

Weak references are not supported though without the Objective-C runtime.

Forwarding implementations

Since CFRetained<T> is a smart pointer, it Derefs to T.

It also forwards the implementation of a bunch of standard library traits such as PartialEq, AsRef, and so on, so that it becomes possible to use e.g. CFRetained<CFString> as-if it was CFString. Note that having a CFString directly is not possible since CoreFoundation objects cannot live on the stack, but instead must reside on the heap, and as such must be accessed behind a pointer or a reference (i.e. &CFString).

Memory layout

This is guaranteed to have the same size and alignment as a pointer to the object, *const T.

Additionally, it participates in the null-pointer optimization, that is, Option<CFRetained<T>> is guaranteed to have the same size as CFRetained<T>.

Implementations

impl<T: ?Sized + Type> CFRetained<T>

pub unsafe fn from_raw(ptr: NonNull<T>) -> Self

Construct a CFRetained from a pointer that already has +1 retain count.

This is useful when you have been given a pointer to a type from some API that returns a retained pointer (i.e. follows the Create rule).

Safety

You must uphold the same requirements as described in CFRetained::retain.

Additionally, you must ensure the given object pointer has +1 retain count.

pub fn into_raw(this: Self) -> NonNull<T>

Consumes the CFRetained, returning a raw pointer with +1 retain count.

After calling this function, the caller is responsible for the memory previously managed by the CFRetained.

This is effectively the opposite of CFRetained::from_raw, see that for more details on when this function is useful.

This is an associated method, and must be called as CFRetained::into_raw(obj).

pub fn as_ptr(this: &Self) -> NonNull<T>

Returns a raw pointer to the type.

The pointer is valid for at least as long as the CFRetained is held.

This is an associated method, and must be called as CFRetained::as_ptr(&obj).

pub unsafe fn cast_unchecked<U: Type>(this: Self) -> CFRetained<U>

Unchecked conversion to another CoreFoundation type.

This is equivalent to an unsafe cast between two pointers, see CFRetained::downcast for a safe alternative.

This is an associated method, and must be called as CFRetained::cast_unchecked(obj).

Safety

You must ensure that the type can be reinterpreted as the given type.

If T is not 'static, you must ensure that U ensures that the data contained by T is kept alive for as long as U lives.

Additionally, you must ensure that any safety invariants that the new type has are upheld.

impl<T: Type> CFRetained<T>

pub fn downcast<U: ConcreteType>(self) -> Result<CFRetained<U>, Self>
where T: 'static,

Attempt to downcast the type to that of type U.

This is the owned variant. Use CFType::downcast_ref if you want to convert a reference type. See also ConcreteType for more details on which types support being converted to.

See CFType::downcast_ref for more details.

Errors

If casting failed, this will return the original back as the Err variant. If you do not care about this, and just want an Option, use .downcast().ok().

pub unsafe fn retain(ptr: NonNull<T>) -> Self

Retain the pointer and construct a CFRetained from it.

This is useful when you have been given a pointer to a type from some API that returns a non-retained reference (i.e. follows the Get rule).

See also Type::retain for a safe alternative when you already have a reference to the type.

Safety

The pointer must be valid as a reference (aligned, dereferenceable and initialized, see the std::ptr module for more information).

You must ensure that if T is non-'static (i.e. has a lifetime parameter), that any data that T may reference lives for at least as long as the return value.

pub fn autorelease_ptr(this: Self) -> *mut T

Autoreleases the CFRetained, returning a pointer.

The object is not immediately released, but will be when the innermost / current autorelease pool is drained.

This is an associated method, and must be called as CFRetained::autorelease_ptr(obj).

Safety

This method is safe to call, but the returned pointer is only guaranteed to be valid until the innermost autorelease pool is drained.