objc2_core_foundation

Struct CFTree

pub struct CFTree { /* private fields */ }

This is the type of a reference to CFTrees.

See also Apple’s documentation

Implementations

impl CFTree

pub unsafe fn new( allocator: Option<&CFAllocator>, context: *const CFTreeContext, ) -> Option<CFRetained<CFTree>>

Creates a new mutable tree.

Parameter allocator: The CFAllocator which should be used to allocate memory for the tree and storage for its children. This parameter may be NULL in which case the current default CFAllocator is used. If this reference is not a valid CFAllocator, the behavior is undefined.

Parameter context: A C pointer to a CFTreeContext structure to be copied and used as the context of the new tree. The info parameter will be retained by the tree if a retain function is provided. If this value is not a valid C pointer to a CFTreeContext structure-sized block of storage, the result is undefined. If the version number of the storage is not a valid CFTreeContext version number, the result is undefined.

Returns: A reference to the new CFTree.

Safety

allocator might not allow None.

context must be a valid pointer.

pub fn parent(&self) -> Option<CFRetained<CFTree>>

Returns the parent of the specified tree.

Parameter tree: The tree to be queried. If this parameter is not a valid CFTree, the behavior is undefined.

Returns: The parent of the tree.

pub fn next_sibling(&self) -> Option<CFRetained<CFTree>>

Returns the sibling after the specified tree in the parent tree’s list.

Parameter tree: The tree to be queried. If this parameter is not a valid CFTree, the behavior is undefined.

Returns: The next sibling of the tree.

pub fn first_child(&self) -> Option<CFRetained<CFTree>>

Returns the first child of the tree.

Parameter tree: The tree to be queried. If this parameter is not a valid CFTree, the behavior is undefined.

Returns: The first child of the tree.

pub unsafe fn context(&self, context: *mut CFTreeContext)

Returns the context of the specified tree.

Parameter tree: The tree to be queried. If this parameter is not a valid CFTree, the behavior is undefined.

Parameter context: A C pointer to a CFTreeContext structure to be filled in with the context of the specified tree. If this value is not a valid C pointer to a CFTreeContext structure-sized block of storage, the result is undefined. If the version number of the storage is not a valid CFTreeContext version number, the result is undefined.

Safety

context must be a valid pointer.

pub fn child_count(&self) -> CFIndex

Returns the number of children of the specified tree.

Parameter tree: The tree to be queried. If this parameter is not a valid CFTree, the behavior is undefined.

Returns: The number of children.

pub fn child_at_index(&self, idx: CFIndex) -> Option<CFRetained<CFTree>>

Returns the nth child of the specified tree.

Parameter tree: The tree to be queried. If this parameter is not a valid CFTree, the behavior is undefined.

Parameter idx: The index of the child tree to be returned. If this parameter is less than zero or greater than the number of children of the tree, the result is undefined.

Returns: A reference to the specified child tree.

pub unsafe fn children(&self, children: *mut *mut CFTree)

Fills the buffer with children from the tree.

Parameter tree: The tree to be queried. If this parameter is not a valid CFTree, the behavior is undefined.

Parameter children: A C array of pointer-sized values to be filled with children from the tree. If this parameter is not a valid pointer to a C array of at least CFTreeGetChildCount() pointers, the behavior is undefined.

Safety

children must be a valid pointer.

pub unsafe fn apply_function_to_children( &self, applier: CFTreeApplierFunction, context: *mut c_void, )

Calls a function once for each child of the tree. Note that the applier only operates one level deep, and does not operate on descendents further removed than the immediate children of the tree.

Parameter tree: The tree to be operated upon. If this parameter is not a valid CFTree, the behavior is undefined.

Parameter applier: The callback function to call once for each child of the given tree. If this parameter is not a pointer to a function of the correct prototype, the behavior is undefined. If there are values in the tree which the applier function does not expect or cannot properly apply to, the behavior is undefined.

Parameter context: A pointer-sized user-defined value, which is passed as the second parameter to the applier function, but is otherwise unused by this function. If the context is not what is expected by the applier function, the behavior is undefined.

Safety

applier must be implemented correctly.

context must be a valid pointer.

pub fn find_root(&self) -> Option<CFRetained<CFTree>>

Returns the root tree of which the specified tree is a descendent.

Parameter tree: The tree to be queried. If this parameter is not a valid CFTree, the behavior is undefined.

Returns: A reference to the root of the tree.

pub unsafe fn set_context(&self, context: *const CFTreeContext)

Replaces the context of a tree. The tree releases its retain on the info of the previous context, and retains the info of the new context.

Parameter tree: The tree to be operated on. If this parameter is not a valid CFTree, the behavior is undefined.

Parameter context: A C pointer to a CFTreeContext structure to be copied and used as the context of the new tree. The info parameter will be retained by the tree if a retain function is provided. If this value is not a valid C pointer to a CFTreeContext structure-sized block of storage, the result is undefined. If the version number of the storage is not a valid CFTreeContext version number, the result is undefined.

Safety

context must be a valid pointer.

pub unsafe fn prepend_child(&self, new_child: Option<&CFTree>)

Adds the newChild to the specified tree as the first in its list of children.

Parameter tree: The tree to be operated on. If this parameter is not a valid CFTree, the behavior is undefined.

Parameter newChild: The child to be added. If this parameter is not a valid CFTree, the behavior is undefined. If this parameter is a tree which is already a child of any tree, the behavior is undefined.

Safety

new_child might not allow None.

pub unsafe fn append_child(&self, new_child: Option<&CFTree>)

Adds the newChild to the specified tree as the last in its list of children.

Parameter tree: The tree to be operated on. If this parameter is not a valid CFTree, the behavior is undefined.

Parameter newChild: The child to be added. If this parameter is not a valid CFTree, the behavior is undefined. If this parameter is a tree which is already a child of any tree, the behavior is undefined.

Safety

new_child might not allow None.

pub unsafe fn insert_sibling(&self, new_sibling: Option<&CFTree>)

Inserts newSibling into the the parent tree’s linked list of children after tree. The newSibling will have the same parent as tree.

Parameter tree: The tree to insert newSibling after. If this parameter is not a valid CFTree, the behavior is undefined. If the tree does not have a parent, the behavior is undefined.

Parameter newSibling: The sibling to be added. If this parameter is not a valid CFTree, the behavior is undefined. If this parameter is a tree which is already a child of any tree, the behavior is undefined.

Safety

new_sibling might not allow None.

pub fn remove(&self)

Removes the tree from its parent.

Parameter tree: The tree to be removed. If this parameter is not a valid CFTree, the behavior is undefined.

pub fn remove_all_children(&self)

Removes all the children of the tree.

Parameter tree: The tree to remove all children from. If this parameter is not a valid CFTree, the behavior is undefined.

pub unsafe fn sort_children( &self, comparator: CFComparatorFunction, context: *mut c_void, )

Sorts the children of the specified tree using the specified comparator function.

Parameter tree: The tree to be operated on. If this parameter is not a valid CFTree, the behavior is undefined.

Parameter comparator: The function with the comparator function type signature which is used in the sort operation to compare children of the tree with the given value. If this parameter is not a pointer to a function of the correct prototype, the the behavior is undefined. The children of the tree are sorted from least to greatest according to this function.

Parameter context: A pointer-sized user-defined value, which is passed as the third parameter to the comparator function, but is otherwise unused by this function. If the context is not what is expected by the comparator function, the behavior is undefined.

Safety

comparator must be implemented correctly.

context must be a valid pointer.