objc2_core_foundation

Struct CFType

pub struct CFType { /* private fields */ }

An instance of a Core Foundation type.

This is meant to be used behind a reference. In the future, this will be defined as an extern type.

All Core Foundation types Deref to this type (it can be considered the “root” type).

See also Apple’s documentation.

Implementations

impl CFType

pub fn downcast_ref<T: ConcreteType>(&self) -> Option<&T>

Attempt to downcast the type to that of type T.

This is the reference-variant. Use CFRetained::downcast if you want to convert a retained type. See also ConcreteType for more details on which types support being converted to.

pub fn retain_count(&self) -> usize

Get the reference count of the object.

This function may be useful for debugging. You normally do not use this function otherwise.

Beware that some things (like CFNumbers, small CFStrings etc.) may not have a normal retain count for optimization purposes, and can return usize::MAX in that case.