objc2_core_foundation

Struct CFURL

pub struct CFURL { /* private fields */ }

Apple’s documentation

This is toll-free bridged with NSURL.

Implementations

impl CFURL

pub unsafe fn with_bytes( allocator: Option<&CFAllocator>, url_bytes: *const u8, length: CFIndex, encoding: CFStringEncoding, base_url: Option<&CFURL>, ) -> Option<CFRetained<CFURL>>

Safety

allocator might not allow None.

url_bytes must be a valid pointer.

base_url might not allow None.

pub fn new_data( allocator: Option<&CFAllocator>, url: Option<&CFURL>, encoding: CFStringEncoding, escape_whitespace: bool, ) -> Option<CFRetained<CFData>>

pub unsafe fn new_absolute_url_with_bytes( alloc: Option<&CFAllocator>, relative_url_bytes: *const u8, length: CFIndex, encoding: CFStringEncoding, base_url: Option<&CFURL>, use_compatibility_mode: bool, ) -> Option<CFRetained<CFURL>>

Safety

alloc might not allow None.

relative_url_bytes must be a valid pointer.

base_url might not allow None.

pub fn with_file_system_path( allocator: Option<&CFAllocator>, file_path: Option<&CFString>, path_style: CFURLPathStyle, is_directory: bool, ) -> Option<CFRetained<CFURL>>

pub unsafe fn from_file_system_representation( allocator: Option<&CFAllocator>, buffer: *const u8, buf_len: CFIndex, is_directory: bool, ) -> Option<CFRetained<CFURL>>

Safety

allocator might not allow None.

buffer must be a valid pointer.

pub fn with_file_system_path_relative_to_base( allocator: Option<&CFAllocator>, file_path: Option<&CFString>, path_style: CFURLPathStyle, is_directory: bool, base_url: Option<&CFURL>, ) -> Option<CFRetained<CFURL>>

pub unsafe fn from_file_system_representation_relative_to_base( allocator: Option<&CFAllocator>, buffer: *const u8, buf_len: CFIndex, is_directory: bool, base_url: Option<&CFURL>, ) -> Option<CFRetained<CFURL>>

Safety

allocator might not allow None.

buffer must be a valid pointer.

base_url might not allow None.

pub unsafe fn file_system_representation( &self, resolve_against_base: bool, buffer: *mut u8, max_buf_len: CFIndex, ) -> bool

Safety

buffer must be a valid pointer.

pub fn absolute_url(&self) -> Option<CFRetained<CFURL>>

pub fn base_url(&self) -> Option<CFRetained<CFURL>>

pub fn can_be_decomposed(&self) -> bool

pub fn scheme(&self) -> Option<CFRetained<CFString>>

pub fn net_location(&self) -> Option<CFRetained<CFString>>

pub fn path(&self) -> Option<CFRetained<CFString>>

pub unsafe fn strict_path( &self, is_absolute: *mut u8, ) -> Option<CFRetained<CFString>>

Safety

is_absolute must be a valid pointer.

pub fn file_system_path( &self, path_style: CFURLPathStyle, ) -> Option<CFRetained<CFString>>

pub fn has_directory_path(&self) -> bool

pub fn resource_specifier(&self) -> Option<CFRetained<CFString>>

pub fn host_name(&self) -> Option<CFRetained<CFString>>

pub fn port_number(&self) -> i32

pub fn user_name(&self) -> Option<CFRetained<CFString>>

pub fn password(&self) -> Option<CFRetained<CFString>>

pub fn parameter_string( &self, characters_to_leave_escaped: Option<&CFString>, ) -> Option<CFRetained<CFString>>

👎Deprecated: The CFURLCopyParameterString function is deprecated. Post deprecation for applications linked with or after the macOS 10.15, and for all iOS, watchOS, and tvOS applications, CFURLCopyParameterString will always return NULL, and the CFURLCopyPath(), CFURLCopyStrictPath(), and CFURLCopyFileSystemPath() functions will return the complete path including the semicolon separator and params component if the URL string contains them.

pub fn query_string( &self, characters_to_leave_escaped: Option<&CFString>, ) -> Option<CFRetained<CFString>>

pub fn fragment( &self, characters_to_leave_escaped: Option<&CFString>, ) -> Option<CFRetained<CFString>>

pub fn last_path_component(&self) -> Option<CFRetained<CFString>>

pub fn path_extension(&self) -> Option<CFRetained<CFString>>

pub fn new_copy_appending_path_component( allocator: Option<&CFAllocator>, url: Option<&CFURL>, path_component: Option<&CFString>, is_directory: bool, ) -> Option<CFRetained<CFURL>>

pub fn new_copy_deleting_last_path_component( allocator: Option<&CFAllocator>, url: Option<&CFURL>, ) -> Option<CFRetained<CFURL>>

pub fn new_copy_appending_path_extension( allocator: Option<&CFAllocator>, url: Option<&CFURL>, extension: Option<&CFString>, ) -> Option<CFRetained<CFURL>>

pub fn new_copy_deleting_path_extension( allocator: Option<&CFAllocator>, url: Option<&CFURL>, ) -> Option<CFRetained<CFURL>>

pub unsafe fn bytes(&self, buffer: *mut u8, buffer_length: CFIndex) -> CFIndex

Safety

buffer must be a valid pointer.

impl CFURL

pub unsafe fn byte_range_for_component( &self, component: CFURLComponentType, range_including_separators: *mut CFRange, ) -> CFRange

Safety

range_including_separators must be a valid pointer.

pub fn new_string_by_replacing_percent_escapes( allocator: Option<&CFAllocator>, original_string: Option<&CFString>, characters_to_leave_escaped: Option<&CFString>, ) -> Option<CFRetained<CFString>>

pub unsafe fn new_string_by_replacing_percent_escapes_using_encoding( allocator: Option<&CFAllocator>, orig_string: Option<&CFString>, chars_to_leave_escaped: Option<&CFString>, encoding: CFStringEncoding, ) -> Option<CFRetained<CFString>>

👎Deprecated: Use [NSString stringByRemovingPercentEncoding] or CFURLCreateStringByReplacingPercentEscapes() instead, which always uses the recommended UTF-8 encoding.

Safety

allocator might not allow None.

orig_string might not allow None.

chars_to_leave_escaped might not allow None.

pub unsafe fn new_string_by_adding_percent_escapes( allocator: Option<&CFAllocator>, original_string: Option<&CFString>, characters_to_leave_unescaped: Option<&CFString>, legal_url_characters_to_be_escaped: Option<&CFString>, encoding: CFStringEncoding, ) -> Option<CFRetained<CFString>>

👎Deprecated: Use [NSString stringByAddingPercentEncodingWithAllowedCharacters:] instead, which always uses the recommended UTF-8 encoding, and which encodes for a specific URL component or subcomponent (since each URL component or subcomponent has different rules for what characters are valid).

Safety

allocator might not allow None.

original_string might not allow None.

characters_to_leave_unescaped might not allow None.

legal_url_characters_to_be_escaped might not allow None.

pub fn is_file_reference_url(&self) -> bool

pub unsafe fn new_file_reference_url( allocator: Option<&CFAllocator>, url: Option<&CFURL>, error: *mut *mut CFError, ) -> Option<CFRetained<CFURL>>

Safety

allocator might not allow None.

url might not allow None.

error must be a valid pointer.

pub unsafe fn new_file_path_url( allocator: Option<&CFAllocator>, url: Option<&CFURL>, error: *mut *mut CFError, ) -> Option<CFRetained<CFURL>>

Safety

allocator might not allow None.

url might not allow None.

error must be a valid pointer.

pub unsafe fn resource_property_for_key( &self, key: Option<&CFString>, property_value_type_ref_ptr: *mut c_void, error: *mut *mut CFError, ) -> bool

Safety

key might not allow None.

property_value_type_ref_ptr must be a valid pointer.

error must be a valid pointer.

pub unsafe fn resource_properties_for_keys( &self, keys: Option<&CFArray>, error: *mut *mut CFError, ) -> Option<CFRetained<CFDictionary>>

Safety

keys generic must be of the correct type.

keys might not allow None.

error must be a valid pointer.

pub unsafe fn set_resource_property_for_key( &self, key: Option<&CFString>, property_value: Option<&CFType>, error: *mut *mut CFError, ) -> bool

Safety

key might not allow None.

property_value should be of the correct type.

property_value might not allow None.

error must be a valid pointer.

pub unsafe fn set_resource_properties_for_keys( &self, keyed_property_values: Option<&CFDictionary>, error: *mut *mut CFError, ) -> bool

Safety

keyed_property_values generic must be of the correct type.

keyed_property_values generic must be of the correct type.

keyed_property_values might not allow None.

error must be a valid pointer.

impl CFURL

pub fn clear_resource_property_cache_for_key(&self, key: Option<&CFString>)

pub fn clear_resource_property_cache(&self)

pub fn set_temporary_resource_property_for_key( &self, key: Option<&CFString>, property_value: Option<&CFType>, )

pub unsafe fn resource_is_reachable(&self, error: *mut *mut CFError) -> bool

Safety

error must be a valid pointer.

impl CFURL

pub unsafe fn new_bookmark_data( allocator: Option<&CFAllocator>, url: Option<&CFURL>, options: CFURLBookmarkCreationOptions, resource_properties_to_include: Option<&CFArray>, relative_to_url: Option<&CFURL>, error: *mut *mut CFError, ) -> Option<CFRetained<CFData>>

Safety

allocator might not allow None.

url might not allow None.

resource_properties_to_include generic must be of the correct type.

resource_properties_to_include might not allow None.

relative_to_url might not allow None.

error must be a valid pointer.

pub unsafe fn new_by_resolving_bookmark_data( allocator: Option<&CFAllocator>, bookmark: Option<&CFData>, options: CFURLBookmarkResolutionOptions, relative_to_url: Option<&CFURL>, resource_properties_to_include: Option<&CFArray>, is_stale: *mut u8, error: *mut *mut CFError, ) -> Option<CFRetained<CFURL>>

Safety

allocator might not allow None.

bookmark might not allow None.

relative_to_url might not allow None.

resource_properties_to_include generic must be of the correct type.

resource_properties_to_include might not allow None.

is_stale must be a valid pointer.

error must be a valid pointer.

pub unsafe fn new_resource_properties_for_keys_from_bookmark_data( allocator: Option<&CFAllocator>, resource_properties_to_return: Option<&CFArray>, bookmark: Option<&CFData>, ) -> Option<CFRetained<CFDictionary>>

Safety

allocator might not allow None.

resource_properties_to_return generic must be of the correct type.

resource_properties_to_return might not allow None.

bookmark might not allow None.

pub unsafe fn new_resource_property_for_key_from_bookmark_data( allocator: Option<&CFAllocator>, resource_property_key: Option<&CFString>, bookmark: Option<&CFData>, ) -> Option<CFRetained<CFType>>

Safety

allocator might not allow None.

resource_property_key might not allow None.

bookmark might not allow None.

pub unsafe fn new_bookmark_data_from_file( allocator: Option<&CFAllocator>, file_url: Option<&CFURL>, error_ref: *mut *mut CFError, ) -> Option<CFRetained<CFData>>

Safety

allocator might not allow None.

file_url might not allow None.

error_ref must be a valid pointer.

pub unsafe fn write_bookmark_data_to_file( bookmark_ref: Option<&CFData>, file_url: Option<&CFURL>, options: CFURLBookmarkFileCreationOptions, error_ref: *mut *mut CFError, ) -> bool

Safety

bookmark_ref might not allow None.

file_url might not allow None.

error_ref must be a valid pointer.

pub unsafe fn new_bookmark_data_from_alias_record( allocator_ref: Option<&CFAllocator>, alias_record_data_ref: Option<&CFData>, ) -> Option<CFRetained<CFData>>

👎Deprecated: The Carbon Alias Manager is deprecated. This function should only be used to convert Carbon AliasRecords to bookmark data.

Safety

allocator_ref might not allow None.

alias_record_data_ref might not allow None.

pub unsafe fn start_accessing_security_scoped_resource(&self) -> bool

pub unsafe fn stop_accessing_security_scoped_resource(&self)

impl CFURL

pub unsafe fn new_data_and_properties_from_resource( alloc: Option<&CFAllocator>, url: Option<&CFURL>, resource_data: *mut *const CFData, properties: *mut *const CFDictionary, desired_properties: Option<&CFArray>, error_code: *mut i32, ) -> bool

👎Deprecated: For resource data, use the CFReadStream API. For file resource properties, use CFURLCopyResourcePropertiesForKeys.

Safety

alloc might not allow None.

url might not allow None.

resource_data must be a valid pointer.

properties must be a valid pointer.

desired_properties generic must be of the correct type.

desired_properties might not allow None.

error_code must be a valid pointer.

pub unsafe fn write_data_and_properties_to_resource( &self, data_to_write: Option<&CFData>, properties_to_write: Option<&CFDictionary>, error_code: *mut i32, ) -> bool

👎Deprecated: For resource data, use the CFWriteStream API. For file resource properties, use CFURLSetResourcePropertiesForKeys.

Safety

data_to_write might not allow None.

properties_to_write generic must be of the correct type.

properties_to_write generic must be of the correct type.

properties_to_write might not allow None.

error_code must be a valid pointer.

pub unsafe fn destroy_resource(&self, error_code: *mut i32) -> bool

👎Deprecated: Use CFURLGetFileSystemRepresentation and removefile(3) instead.

Safety

error_code must be a valid pointer.

pub unsafe fn new_property_from_resource( alloc: Option<&CFAllocator>, url: Option<&CFURL>, property: Option<&CFString>, error_code: *mut i32, ) -> Option<CFRetained<CFType>>

👎Deprecated: For file resource properties, use CFURLCopyResourcePropertyForKey.

Safety

alloc might not allow None.

url might not allow None.

property might not allow None.

error_code must be a valid pointer.

impl CFURL

Path conversion.

pub fn from_file_path<P: AsRef<Path>>(path: P) -> Option<CFRetained<CFURL>>

Create a file url from a Path.

This is useful because a lot of CoreFoundation APIs use CFURL to represent file-system paths as well.

Non-unicode parts of the URL will be percent-encoded, and the url will have the scheme file://.

If the path is relative, it will be considered relative to the current directory.

Returns None when given an invalid path (such as a path containing interior NUL bytes). The exact checks are not guaranteed.

Examples

use std::path::Path;
use objc2_core_foundation::CFURL;

// Absolute paths work as you'd expect.
let url = CFURL::from_file_path("/tmp/file.txt").unwrap();
assert_eq!(url.to_file_path().unwrap(), Path::new("/tmp/file.txt"));

// Relative paths are relative to the current directory.
let url = CFURL::from_file_path("foo.txt").unwrap();
assert_eq!(url.to_file_path().unwrap(), std::env::current_dir().unwrap().join("foo.txt"));

// Some invalid paths return `None`.
assert!(CFURL::from_file_path("").is_none());
// Another example of an invalid path containing interior NUL bytes.
assert!(CFURL::from_file_path("/a/\0a").is_none());

// Trailing NUL bytes are stripped.
// NOTE: This only seems to work on some versions of CoreFoundation.
let url = CFURL::from_file_path("/a\0\0").unwrap();
assert_eq!(url.to_file_path().unwrap(), Path::new("/a"));

pub fn from_directory_path<P: AsRef<Path>>(path: P) -> Option<CFRetained<CFURL>>

Create a directory url from a Path.

This differs from from_file_path in that the path is treated as a directory, which means that other normalization rules are applied to it (to make it end with a /).

Examples

use std::path::Path;
use objc2_core_foundation::CFURL;

// Directory paths get trailing slashes appended
let url = CFURL::from_directory_path("/Library").unwrap();
assert_eq!(url.to_file_path().unwrap(), Path::new("/Library/"));

// Unless they already have them.
let url = CFURL::from_directory_path("/Library/").unwrap();
assert_eq!(url.to_file_path().unwrap(), Path::new("/Library/"));

// Similarly for relative paths.
let url = CFURL::from_directory_path("foo").unwrap();
assert_eq!(url.to_file_path().unwrap(), std::env::current_dir().unwrap().join("foo/"));

// Various dots may be stripped.
let url = CFURL::from_directory_path("/Library/././.").unwrap();
assert_eq!(url.to_file_path().unwrap(), Path::new("/Library/"));

// Though of course not if they have semantic meaning.
let url = CFURL::from_directory_path("/Library/..").unwrap();
assert_eq!(url.to_file_path().unwrap(), Path::new("/Library/.."));

pub fn to_file_path(&self) -> Option<PathBuf>

Extract the path part of the URL as a PathBuf.

This will return a path regardless of whether the scheme is file://. It is the responsibility of the caller to ensure that the URL is valid to use as a file URL.

Compatibility note

This currently does not work for non-unicode paths (which are fairly rare on macOS since HFS+ was been superseded by APFS).

This also currently always returns absolute paths (it converts relative URL paths to absolute), but that may change in the future.

Examples

use std::path::Path;
use objc2_core_foundation::{CFURL, CFString};

let url = CFURL::from_string(None, &CFString::from_str("file:///tmp/foo.txt"), None).unwrap();
assert_eq!(url.to_file_path().unwrap(), Path::new("/tmp/foo.txt"));

See also the examples in from_file_path.

impl CFURL

String conversion.

pub fn from_string( allocator: Option<&CFAllocator>, url_string: &CFString, base_url: Option<&CFURL>, ) -> Option<CFRetained<Self>>

Create an URL from a CFString.

Returns None if the URL is considered invalid by CoreFoundation. The exact details of which strings are invalid URLs are considered an implementation detail.

Note in particular that not all strings that the URL spec considers invalid are considered invalid by CoreFoundation too. If you need spec-compliant parsing, consider the url crate instead.

Examples

Construct and inspect a CFURL.

use objc2_core_foundation::{
    CFString, CFURL, CFURLCopyHostName, CFURLCopyScheme, CFURLCopyPath,
};

let url = CFURL::from_string(None, &CFString::from_str("http://example.com/foo"), None).unwrap();
assert_eq!(url.string().to_string(), "http://example.com/foo");
assert_eq!(CFURLCopyScheme(&url).unwrap().to_string(), "http");
assert_eq!(CFURLCopyHostName(&url).unwrap().to_string(), "example.com");
assert_eq!(CFURLCopyPath(&url).unwrap().to_string(), "/foo");

Fail parsing certain strings.

use objc2_core_foundation::{CFString, CFURL};

// Percent-encoding needs two characters.
assert_eq!(CFURL::from_string(None, &CFString::from_str("http://example.com/%A"), None), None);

// Two hash-characters is disallowed.
assert_eq!(CFURL::from_string(None, &CFString::from_str("http://example.com/abc#a#b"), None), None);

pub unsafe fn from_str_unchecked(s: &str) -> Option<CFRetained<Self>>

Create an URL from a string without checking it for validity.

Returns None on some OS versions when the string contains interior NUL bytes.

Safety

The URL must be valid.

Note that it is unclear whether this is actually a safety requirement, or simply a correctness requirement. So we conservatively mark this function as unsafe.

pub fn string(&self) -> CFRetained<CFString>

Get the string-representation of the URL.

The string may be overly sanitized (percent-encoded), do not rely on this returning exactly the same string as was passed in from_string.