objc2_core_foundation

Trait Type

pub unsafe trait Type {
    // Provided method
    fn retain(&self) -> CFRetained<Self> 
       where Self: Sized { ... }
}

A CoreFoundation-like type.

This trait is implemented for all CoreFoundation-like types (i.e. both types in this crate like CFString, but also types from other frameworks like CGColor from CoreGraphics).

This trait allows the type to be used in CFRetained.

You should not need to implement this yourself.

Safety

The type must be a CoreFoundation-like type.

The type must be safe to retain/release using CFRetain and CFRelease.

Provided Methods

fn retain(&self) -> CFRetained<Self>
where Self: Sized,

Increment the reference count of the receiver.

This extends the duration in which the receiver is alive by detaching it from the lifetime information carried by the reference.

This is similar to using Clone on CFRetained<Self>, with the addition that it can be used on a plain reference.