objc2_core_foundation

Trait Type

pub unsafe trait Type {
    // Provided methods
    fn retain(&self) -> CFRetained<Self> 
       where Self: Sized { ... }
    fn as_concrete_TypeRef(&self) -> &Self { ... }
    unsafe fn wrap_under_get_rule(ptr: *const Self) -> CFRetained<Self> 
       where Self: Sized { ... }
    fn as_CFTypeRef(&self) -> &CFType
       where Self: AsRef<CFType> { ... }
    unsafe fn wrap_under_create_rule(ptr: *const Self) -> CFRetained<Self> 
       where Self: Sized { ... }
}

A CoreFoundation-like type.

This trait is implemented for all CoreFoundation-like types (i.e. both types in this crate like CFString, but also types from other frameworks like CGColor from CoreGraphics).

This trait allows the type to be used in CFRetained.

You should not need to implement this yourself.

Safety

The type must be a CoreFoundation-like type.

The type must be safe to retain/release using CFRetain and CFRelease.

Provided Methods

fn retain(&self) -> CFRetained<Self>
where Self: Sized,

Increment the reference count of the receiver.

This extends the duration in which the receiver is alive by detaching it from the lifetime information carried by the reference.

This is similar to using Clone on CFRetained<Self>, with the addition that it can be used on a plain reference.

fn as_concrete_TypeRef(&self) -> &Self

๐Deprecated: this is redundant

Helper for easier transition from the core-foundation crate.

unsafe fn wrap_under_get_rule(ptr: *const Self) -> CFRetained<Self>
where Self: Sized,

๐Deprecated: use CFRetained::retain

Helper for easier transition from the core-foundation crate.

Safety

Same as CFRetained::retain.

fn as_CFTypeRef(&self) -> &CFType
where Self: AsRef<CFType>,

๐Deprecated: this is redundant (CF types deref to CFType)

Helper for easier transition from the core-foundation crate.

unsafe fn wrap_under_create_rule(ptr: *const Self) -> CFRetained<Self>
where Self: Sized,

๐Deprecated: use CFRetained::from_raw

Helper for easier transition from the core-foundation crate.

Safety

Same as CFRetained::from_raw.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.